% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addMemory}
\alias{addMemory}
\alias{addMemory.memoryModeler}
\title{add new data into memoryModeler}
\usage{
addMemory(object, data)

\method{addMemory}{memoryModeler}(object, data)
}
\arguments{
\item{object}{\code{\link{modeler}} object to fit to the data.}

\item{data}{Dataframe containing the data to use for the training}
}
\value{
memoryModeler object with new data loaded into list of data sets.
}
\description{
Add new data to memoryModeler and remove oldest data if necessary
}
\examples{
#Generate data
set.seed(273)
data.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 1000, generate_random_weights = FALSE)
data2.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 100, generate_random_weights = FALSE)
data3.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 50, generate_random_weights = FALSE)

#Define train/test partitions
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

#Define base modeler
target <- "target"# Regression
regressors <- setdiff(names(data.dt), target)
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)
base_modeler <- rangerRegressorModeler(params = params,
                                       target = target,
                                       regressors = regressors)

#Instantiate memoryModeler
#memoryModeler parameters
data <- list(data2.dt)
maxData <- 3L
modeler <- memoryRegressorModeler(data = data,
                                  modeler = base_modeler,
                                  maxData = maxData)

#Add new data to modeler
length(modeler$data)
modeler <- addMemory(object = modeler,
data = data3.dt)
length(modeler$data)

#If data list exceeds maxData length, the oldest dataset is deleted
modeler <- addMemory(object = modeler,
data = data3.dt)
modeler <- addMemory(object = modeler,
data = data3.dt)

length(modeler$data)

}
\seealso{
\code{\link{memoryRegressorModeler}}

\code{\link{cleanMemory}}

\code{\link{updateMemory}}
}
