% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assignExternalAttributes}
\alias{assignExternalAttributes}
\alias{assignExternalAttributes.modeler}
\alias{assignExternalAttributes.prePredRegressorModeler}
\alias{assignExternalAttributes.classifRegRegressorModeler}
\alias{assignExternalAttributes.benchModeler}
\alias{assignExternalAttributes.list}
\title{Method to assign required attributes to a preproObject.}
\usage{
assignExternalAttributes(modeler, object)

\method{assignExternalAttributes}{modeler}(modeler, object)

\method{assignExternalAttributes}{prePredRegressorModeler}(modeler, object)

\method{assignExternalAttributes}{classifRegRegressorModeler}(modeler, object)

\method{assignExternalAttributes}{benchModeler}(modeler, object)

\method{assignExternalAttributes}{list}(modeler, object)
}
\arguments{
\item{modeler}{A modeler object with attributes relevant to preprocessing,
such as target(s), regressors and weightsColumn. It can also be a list with
target, regressors and weightsColumn components, in case the user wants to
preprocess data independently from any specific modeler.}

\item{object}{A preproObject.}
}
\value{
preproObject with target, regressors and weightsColumn attributes.
}
\description{
Assigns target, regressors and weightsColumn to a preproObject.
Additionally, defines which columns to filter when training and predicting.

This method runs automatically when instantiating a preproModeler. The user
should avoid running it unless they want to preprocess data outside a
preproModeler.
}
\examples{
#Generate data
data.dt <- generateRegressionData(n_samples = 1000,
                                  generate_random_weights = TRUE)
head(data.dt)

#Define relevant variables
target <- "regression_target"
weightsColumn <- "weights"
regressors <- setdiff(colnames(data.dt), c(target, weightsColumn))

modeler <- list("target" = target,
                "regressors" = regressors,
                "weightsColumn" = weightsColumn)

#Define preproObject
#Preprocessing parameters
prepro_params <- list(pretrained_min_max = NULL)

#Define preproObject
prepro_object <- preproObject(preproFunction = minMaxScalerPreproFunction,
                              params = prepro_params)

#Assign relevant metadata to preproObject
prepro_object <- assignExternalAttributes(modeler = modeler,
                                          object = prepro_object)

#Call getPreproData to obtain preproObject with preprocessed data
prepro_object <- getPreproData(object = prepro_object, data = data.dt)

#data is saved within prepro_object
head(prepro_object$data)


#Alternatively, assignExternalAttributes can work with any modeler
params <- list(
  "learning_rate" = 0.07,
  "nrounds" = 10,
  "objective" = "regression",
  "weight_column" = "weights"
)

modeler <- LGBRegressorModeler(params = params,
                               target = target,
                               regressors = regressors)

#Define preproObject
prepro_object <- preproObject(preproFunction = minMaxScalerPreproFunction,
                              params = prepro_params)

#Assign relevant metadata to preproObject
prepro_object <- assignExternalAttributes(modeler = modeler,
                                          object = prepro_object)

#As it can be seen, preproObject now has the matching attributes provided
#by the modeler (note that the weightsColumn attribute is now NULL)
print(prepro_object)

#Call getPreproData to obtain preproObject with preprocessed data
prepro_object <- getPreproData(object = prepro_object, data = data.dt)

#data is saved within prepro_object
#note that our data has a weights column, and the previous modeler does
#have a weightsColumn attribute (since it was defined in its parameters),
#so the weights column was transformed.
head(prepro_object$data)

}
\seealso{
\code{\link{preproRegressorModeler}}

\code{\link{preproClassifierModeler}}

\code{\link{preproObject}}

\code{\link{getPreproData}}
}
