% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorModeler.R
\name{benchRegressorModeler}
\alias{benchRegressorModeler}
\alias{benchRegressorModeler.list}
\title{Class constructor for benchRegressorModeler}
\usage{
benchRegressorModeler(submodels, benchmarks, models, optWeights = "uniform")

\method{benchRegressorModeler}{list}(submodels, benchmarks, optWeights = "uniform")
}
\arguments{
\item{submodels}{List of modelers to be used when predicting.}

\item{benchmarks}{String or character vector of linear restrictions to be
taken into account by \code{benchRegressorModeler}. Each restriction must
be a linear equality or inequality of the form
\code{LHS == RHS}, \code{LHS >= RHS}, or \code{LHS <= RHS}, where \code{LHS}
is a sum or difference of terms. A term can be an R identifier
(\code{X}), a signed identifier (\code{-X}), or a numeric coefficient
multiplied by an identifier (\code{2 * X}, \code{-2.5 * X}). The
right-hand side must be numeric. For example:
\code{"2 * X - Y + 0.5 * Z >= -10"}.}

\item{optWeights}{Parameter indicating the type of optional weights: "uniform" or "MSE". "uniform" is set as default.}
}
\value{
A benchRegressorModeler object.
}
\description{
A multivariate modeler where the target variables must accomplish a certain restriction. It requires multiple modelers to be instantiated.
}
\examples{
#Generate data
original_data.dt <- generateRegressionData(n_samples = 1000, generate_random_weights = FALSE)

#Define new variables in the data that have a specific restriction
#In our case: regression_target = A + B
data.dt <- data.table::copy(original_data.dt)
data.dt[, A := round(runif(dim(data.dt)[1], min = 0, max = 5))]
data.dt[, B := regression_target - A]

#Define train/test partitions
set.seed(273)
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

#Define submodeler parameters
#Note that none of the modelers sees any other target variable besides than their own.
params_b1 <- list(
  "num.trees" = 20,
  "max.depth" = 20
)
target_b1 <- "regression_target"
regressors_b1 <- setdiff(colnames(original_data.dt), target_b1)

params_b2 <- list(
  "num.trees" = 30,
  "max.depth" = 10
)
target_b2 <- "A"
regressors_b2 <- setdiff(colnames(original_data.dt), c(target_b2, target_b1))

params_b3 <- list(
  "num.trees" = 40,
  "max.depth" = 15
)
target_b3 <- "B"
regressors_b3 <- setdiff(colnames(original_data.dt), c(target_b3, target_b1))

modeler1 <- rangerRegressorModeler(params = params_b1, target = target_b1, regressors = regressors_b1)

modeler2 <- rangerRegressorModeler(params = params_b2, target = target_b2, regressors = regressors_b2)

modeler3 <- rangerRegressorModeler(params = params_b3, target = target_b3, regressors = regressors_b3)

#Define bench regressor modeler
b_modeler <- benchRegressorModeler(submodels = list(modeler1,modeler2,modeler3),
                                   benchmarks = "A+B-regression_target==0")

#Train and evaluate modeler
b_modeler <- train(b_modeler, data = data.dt, subSet = train.keys)
predictions <- getPredictions(b_modeler, data = data.dt, subSet = test.keys)
cf <- MSECostFunction()
evaluate(b_modeler, data = data.dt, cost_function = cf, subSet = train.keys)
evaluate(b_modeler, data = data.dt, cost_function = cf, subSet = test.keys)

}
\seealso{
\code{\link{regressorModeler}}
}
