% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMapping.R
\name{binaryThresholdClass}
\alias{binaryThresholdClass}
\title{Binary threshold class mapping.}
\usage{
binaryThresholdClass(threshold = 0.5, positive_class = NULL)
}
\arguments{
\item{threshold}{A threshold used to pick the positive class (if it's below that score, the negative class is returned instead).}

\item{positive_class}{The column name of the positive class.}
}
\value{
A classMapping object
}
\description{
Implementation of a binary threshold class mapping algorithm.
When given a score matrix with two columns (a positive and a negative class), the \code{\link{getClassMapping}} function will output an index array indicating which class was chosen.
The expected use of this mapping involves situations in which we require a higher (or lower) score threshold for the positive class to be chosen, rather than simply being larger than 0.5.
}
\examples{

#Define class mapping (check getClassMapping and getClasses documentation to see a more complete example)
cls_mapping <- binaryThresholdClass(threshold = 0.6, positive_class = NULL)


}
\seealso{
\code{\link{getClassMapping}}

\code{\link{getClasses}}

\code{\link{maxScoreClass}}
}
