% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{bootstrapSampler}
\alias{bootstrapSampler}
\title{Bootstrap sampler}
\usage{
bootstrapSampler(key)
}
\arguments{
\item{key}{Vector of indices of the data entries (rows) to sample from.}
}
\value{
A bootstrapSampler object.
}
\description{
Bootstrap sampler implementation. Given a list or array of indices,
a sample with replacement is performed (its size is set to match the key size parameter).
}
\examples{
set.seed(273)
#Define data and key
data.dt <- generateStratifiedDomainEstimateData(n_samples = 10)
key <- 1:nrow(data.dt)
head(data.dt)
smplr <- bootstrapSampler(key)
s1 <- getSample(smplr)
print(data.dt[s1])

}
\seealso{
\code{\link{sampler}}

\code{\link{srsSampler}}

\code{\link{ssSampler}}

\code{\link{cvSampler}}

\code{\link{fullSampler}}
}
