% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierModeler.R
\name{catboostClassifierModeler}
\alias{catboostClassifierModeler}
\alias{catboostClassifierModeler.list}
\title{Class constructor for catboostClassifierModeler}
\usage{
catboostClassifierModeler(
  params,
  target,
  regressors,
  val_data,
  positive_class = NULL,
  ntree_start = 0,
  ntree_end = 0,
  thread_count = -1
)

\method{catboostClassifierModeler}{list}(
  params,
  target,
  regressors,
  val_data = NULL,
  positive_class = NULL,
  ntree_start = 0,
  ntree_end = 0,
  thread_count = -1
)
}
\arguments{
\item{params}{List of parameters to be used for training a catboost model.
For classification, the available loss_function parameters are:
\itemize{
  \item Logloss (Binary)
  \item CrossEntropy (Binary)
  \item MultiClass
  \item MultiClassOneVsAll
}}

\item{target}{String indicating the name of the target variable in the data.}

\item{regressors}{List or vector of strings indicating the regressor column names in the data.}

\item{val_data}{Validation dataset to be used when training.}

\item{positive_class}{String indicating the name of the positive class, only
used when params$loss_function is Logloss or CrossEntropy. Note that if this
parameter is not provided, it will be assumed that the target variable levels
are the negative class and the positive class, in that order.}

\item{ntree_start}{Additional parameter for catboost's predict method.}

\item{ntree_end}{Additional parameter for catboost's predict method.}

\item{thread_count}{Additional parameter for catboost's predict method.}
}
\value{
A catboostClassifierModeler object.
}
\description{
Catboost implementation in rmlutils for classification tasks.
}
\examples{
#Generate data
data.dt <- generateBinaryClassificationData(n_samples = 1000, generate_random_weights = FALSE)
#data.dt <- generateMulticlassClassificationData(n_samples = 1000, generate_random_weights = FALSE)

#Define train/test partitions
set.seed(273)
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

target <- "classification_target"
regressors <- setdiff(colnames(data.dt), target)

#Define and train modeler
params <- list(
  #loss_function = 'MultiClass',
  #loss_function = 'MultiClassOneVsAll',
  loss_function = 'Logloss',#Binary
  #loss_function = 'CrossEntropy',#Binary
  iterations = 100,
  metric_period=10)

modeler <- catboostClassifierModeler(params = params,
                                     target = target,
                                     regressors = regressors,
                                     positive_class = "positive")

modeler <- train(object = modeler, data = data.dt, subSet = train.keys)

#Get predictions and classes
preds <- getPredictions(object = modeler,
                        data = data.dt,
                        subSet = train.keys)

clss <- getClasses(object = modeler,
                   data = data.dt,
                   subSet = train.keys,
                   class_mapping = maxScoreClass())

#cf <- confusionMatrixCostFunction()
cf <- accuracyCostFunction()

#Obtain cost with getCost (this way we see the confusion matrix)
getCost(object = cf,
        data = data.dt,
        modeler = modeler,
        fSample = 1:nrow(data.dt),
        cSample = train.keys)

#Obtain cost with evaluate
evaluate(object = modeler,
         data = data.dt,
         cost_function = cf,
         subSet = test.keys)


}
\seealso{
\code{\link{classifierModeler}}
}
