% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorModeler.R
\name{catboostRegressorModeler}
\alias{catboostRegressorModeler}
\alias{catboostRegressorModeler.list}
\title{Class constructor for catboostRegressorModeler}
\usage{
catboostRegressorModeler(
  params,
  target,
  regressors,
  val_data,
  ntree_start = 0,
  ntree_end = 0,
  thread_count = -1
)

\method{catboostRegressorModeler}{list}(
  params,
  target,
  regressors,
  val_data = NULL,
  ntree_start = 0,
  ntree_end = 0,
  thread_count = -1
)
}
\arguments{
\item{params}{List of parameters to be used for training a catboost model.}

\item{target}{String indicating the name of the target variable in the data.}

\item{regressors}{List or vector of strings indicating the regressor column names in the data.}

\item{val_data}{Validation dataset to be used when training.}

\item{ntree_start}{Additional parameter for catboost's predict method.}

\item{ntree_end}{Additional parameter for catboost's predict method.}

\item{thread_count}{Additional parameter for catboost's predict method.}
}
\value{
A catboostRegressorModeler object.
}
\description{
Catboost implementation in rmlutils for regression tasks.
}
\examples{
#Generate data
data.dt <- generateRegressionData(n_samples = 1000, generate_random_weights = FALSE)

#Define train/test partitions
set.seed(273)
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

#Define modeler
params <- list(loss_function = 'RMSE',
               iterations = 100,
               metric_period=10)
target <- "regression_target"
regressors <- setdiff(colnames(data.dt), target)

modeler <- catboostRegressorModeler(params = params, target = target, regressors = regressors)

#Train, predict and evaluate
modeler <- train(object = modeler, data = data.dt, subSet = train.keys)

preds <- getPredictions(modeler, data = data.dt, subSet = train.keys)

cf <- r2CostFunction()

evaluate(object = modeler, data = data.dt, cost_function = cf, subSet = train.keys)

}
\seealso{
\code{\link{regressorModeler}}

\code{\link{catboostClassifierModeler}}
}
