% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierModeler.R
\name{classifierModeler}
\alias{classifierModeler}
\title{Class constructors for classification modeler classes.}
\value{
A classifierModeler object.
}
\description{
Creates a modeler object as specified.

The modeler classes contain different types of machine learning models to
provide a common interface for all of them.
}
\section{Current implementations}{


\strong{Simple modelers:}
\itemize{
\item \code{\link{H2OClassifierModeler}}
\item \code{\link{LGBClassifierModeler}}
\item \code{\link{rangerClassifierModeler}}
\item \code{\link{catboostClassifierModeler}}
\item \code{\link{miceClassifierModeler}}
}

\strong{Complex modelers:}
\itemize{
\item \code{\link{preproClassifierModeler}}
}
}

