% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorModeler.R
\name{classifRegRegressorModeler}
\alias{classifRegRegressorModeler}
\alias{classifRegRegressorModeler.modeler}
\title{Class constructor for classifRegRegressorModeler}
\usage{
classifRegRegressorModeler(
  classifModeler,
  regModeler,
  specialCats,
  minRows = 100,
  epsilon = 0,
  class_mapping = maxScoreClass()
)

\method{classifRegRegressorModeler}{modeler}(
  classifModeler,
  regModeler,
  specialCats,
  minRows = 100,
  epsilon = 0,
  class_mapping = maxScoreClass()
)
}
\arguments{
\item{classifModeler}{A classifierModeler to perform a preliminary
classification task.}

\item{regModeler}{A regressorModeler to perform a regression task after the
preliminary classification.}

\item{specialCats}{Special numbers in the target variable, that is, the ones
that should be classified in the first step.}

\item{minRows}{Minimum amount of rows to train the regressorModeler. If this
number is not reached after training the classifierModeler, only an average
will be outputted.}

\item{epsilon}{Tolerance factor. If any target variable value is closer to a
specialCats than this amount, then it gets replaced by the specialCats value.}

\item{class_mapping}{Class mapping algorithm to be used by classifModeler.
Set to maxScoreClass by default.}
}
\value{
A classifRegRegressorModeler object.
}
\description{
A two-step complex modeler. Useful when there are large amounts
of repeated numbers in a numeric target. It works by first performing a
classification task for the repeated numbers.
Then, a regressorModeler is used to predict the remaining samples
(as long as it had a minimum amount of rows to train).
}
\examples{
#Generate data
set.seed(273)
data.dt <- generateRegressionData(n_samples = 1000,
                                  generate_random_weights = FALSE)

# Convert a reasonable number of registers to zero
data.dt[regression_target < 0, regression_target := 0]
table(data.dt[, regression_target == 0])
#We will apply a classifReg modeler to first predict whether the result is 0
#or not, and then automatically apply regression over the remaining data.
#This complete process is internally processed by the classifReg modeler.

#Define train/test partitions
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

#Target variable name is the same for both modelers
target <- "regression_target"
#Regressors could be different, but it is not necessary in this case.
regressors <- setdiff(names(data.dt), target)

#Define a classifier modeler
params_cls <- list(
  "num.trees" = 50,
  "max.depth" = 20
)

cls_modeler <- rangerClassifierModeler(params = params_cls,
                                       target = target,
                                       regressors = regressors)
#Define a regressor modeler
params_reg <- list(
  "learning_rate" = 0.07,
  "nrounds" = 100,
  "objective" = "regression"
)
reg_modeler <- LGBRegressorModeler(params = params_reg,
                                   target = target,
                                   regressors = regressors)

#Define classifReg modeler
modeler <- classifRegRegressorModeler(classifModeler = cls_modeler,
                                      regModeler = reg_modeler,
                                      specialCats = 0)

modeler <- train(object = modeler, data = data.dt, subSet = train.keys)
#Note that classifRegModeler introduces NAs internally while predicting but does
#not alter the proper output

#Define cost function
cf <- r2CostFunction()

#Evaluate using getCost
getCost(object = cf,
        data = data.dt,
        modeler = modeler,
        fSample = data.keys,
        cSample = test.keys)

#Evaluate using evaluate function (only complete modeler)
evaluate(modeler,
         data = data.dt,
         cost_function = cf,
         subSet = test.keys)

#Evaluate internal modelers separately
cf2 <- list("classifModeler" = confusionMatrixCostFunction(),
            "regModeler" = r2CostFunction())
evaluate(object = modeler,
         data = data.dt,
         cost_function = cf2,
         subSet = train.keys)


}
\seealso{
\code{\link{regressorModeler}}

\code{\link{classifierModeler}}
}
