% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMapping.R
\name{classMapping}
\alias{classMapping}
\title{Class mapping from model scores}
\description{
A collection of methods that map a set of \code{\link{classifierModeler}} prediction scores to a set of classes.
}
\section{Current implementations}{


\itemize{
\item \code{\link{maxScoreClass}}
\item \code{\link{binaryThresholdClass}}
}
}

