% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanMemory}
\alias{cleanMemory}
\alias{cleanMemory.memoryModeler}
\title{clean data from memoryModeler}
\usage{
cleanMemory(object, index = 1:object$maxData)

\method{cleanMemory}{memoryModeler}(object, index = 1:object$maxData)
}
\arguments{
\item{object}{\code{\link{modeler}} object to fit to the data.}

\item{index}{vector containing the positions of the data sets to be removed.}
}
\value{
memoryModeler object without all or some data sets.
}
\description{
clean a certain number of data sets from memoryModeler. By default all data sets are removed
}
\examples{
#Generate data
set.seed(273)
data.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 1000, generate_random_weights = FALSE)
data2.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 100, generate_random_weights = FALSE)
data3.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 50, generate_random_weights = FALSE)

#Define train/test partitions
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

#Define base modeler
target <- "target"# Regression
regressors <- setdiff(names(data.dt), target)
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)
base_modeler <- rangerRegressorModeler(params = params,
                                       target = target,
                                       regressors = regressors)

#Instantiate memoryModeler
#memoryModeler parameters
data <- list(data2.dt)
maxData <- 3L
modeler <- memoryRegressorModeler(data = data,
                                  modeler = base_modeler,
                                  maxData = maxData)
#cleanMemory can delete all data (if index is left empty) or any specific dataset
#Let us delete all data in this example
length(modeler$data)
modeler <- cleanMemory(object = modeler)
length(modeler$data)

}
\seealso{
\code{\link{memoryRegressorModeler}}

\code{\link{addMemory}}

\code{\link{updateMemory}}
}
