% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionedSampler.R
\name{conditionedSampler}
\alias{conditionedSampler}
\alias{conditionedSampler.srsSampler}
\alias{conditionedSampler.ssSampler}
\title{Conditioned sampler.}
\usage{
conditionedSampler(object, sampleKeys)

\method{conditionedSampler}{srsSampler}(object, sampleKeys)

\method{conditionedSampler}{ssSampler}(object, sampleKeys)
}
\arguments{
\item{object}{sampler object}

\item{sampleKeys}{list of keys obtained from a sample.}
}
\value{
A sampler.
}
\description{
Generates the conditioned sampler from a sampler and a subsample.

The subsampler is assumed to be independent on the sample. Otherwise the
subsampler would be needed to estimate the conditioned sampler.
}
\examples{
smplr1 <- srsSampler(1:20, 5)
sampleKeys1 <- c(2,7,8,10)
cSampler1 <- conditionedSampler(object = smplr1, sampleKeys = sampleKeys1)
cSampler1
#returns a smplr1 object with keys 1:20 except 2,7,8,10
#notice that it changes the sample size to 5 - lenght(sampleKeys)

smplr2 <- ssSampler(1:20, rep(c("ODDS", "EVENS"), 10), c(EVENS = 8, ODDS = 2))
sampleKeys2 <- c(2,4,6,8,10,11)
cSampler2 <- conditionedSampler(object = smplr2, sampleKeys = sampleKeys2)
cSampler2
#returns a smplr2 object with keys 1:20 except 2,4,6,8,10,11
#notice that it changes the sample size to size
#EVENS = 8 - length(2,4,6,8,10)
#ODDS = 2 - length(11)



}
