% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costFunctions.R
\name{confusionMatrixCostFunction}
\alias{confusionMatrixCostFunction}
\title{Confusion matrix cost function}
\usage{
confusionMatrixCostFunction(
  weightsColumn = "",
  class_map = maxScoreClass(),
  normalize = NULL,
  class_names = NULL
)
}
\arguments{
\item{weightsColumn}{Name of the weights column in the data}

\item{class_map}{\code{\link{classMapping}} object to use (\code{\link{maxScoreClass}} by default)}

\item{normalize}{String indicating how to normalize the matrix. Possible values are:

\itemize{
  \item "true": normalize by true labels (rows)
  \item "pred" : normalize by predicted labels (columns)
  \item "all": normalize by total samples
  \item NULL: no normalization (default)
}}

\item{class_names}{Vector containing the levels of the categorical variable. This only affects the output of the getReducedCost method.}
}
\value{
Named list with all relevant cost metrics.
}
\description{
A costFunction object that outputs a confusion matrix as a performance metric for a \code{\link{classifierModeler}}.
Like any other cost function, it must be used together with the \code{\link{getCost}} function to output its metric.
}
\examples{
#Generate data
data.dt <- generateMulticlassClassificationData(n_samples = 1000)
#Instantiate cost function
cf <- confusionMatrixCostFunction(weightsColumn = "",
                                  class_map = maxScoreClass(),
                                  normalize = NULL,
                                  class_names = levels(data.dt[["classification_target"]]))

#Train simple modeler
target <- "classification_target"
regressors <- setdiff(names(data.dt), target)
params <- list(
  "learning_rate" = 0.07,
  "nrounds" = 10,#Avoid warning
  "objective" = "multiclass",
  "num_class" = length(unique(data.dt[, classification_target]))
)

classifier <- LGBClassifierModeler(params = params, target = target, regressors = regressors)
classifier <- train(classifier, data = data.dt)

#Evaluate with getCost
cost <- getCost(object = cf, data = data.dt, modeler = classifier, cSample = 1:nrow(data.dt))
print(cost)

}
\seealso{
\code{\link{costFunction}}

\code{\link{getCost}}
}
