% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costFunctions.R
\name{costFunction}
\alias{costFunction}
\title{A collection of class constructors for several cost functions.}
\description{
Creates a costFunction object as specified.

The cost function classes contain different types of cost functions
to provide a commom interface for all of them.
}
\section{Current implementations}{


\strong{Regression:}
\itemize{
\item \code{\link{r2CostFunction}}
\item \code{\link{squareBiasCostFunction}}
\item \code{\link{MSECostFunction}}
\item \code{\link{groupMSECostFunction}}
\item \code{\link{groupTotalErrorCostFunction}}
\item \code{\link{MCSubRBMSECostFunction}}
\item \code{\link{holdoutTemporalCostFunction}}
}

\strong{Classification:}
\itemize{
\item \code{\link{confusionMatrixCostFunction}}
\item \code{\link{accuracyCostFunction}}
\item \code{\link{precisionCostFunction}}
\item \code{\link{recallCostFunction}}
\item \code{\link{f1ScoreCostFunction}}
\item \code{\link{ROCAUCCostFunction}}
}
}

