% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidate.R
\name{crossValidate}
\alias{crossValidate}
\alias{crossValidate.modeler}
\title{Model cross validation.}
\usage{
crossValidate(modeler, data, CVSampler, costFunction, times = 1)

\method{crossValidate}{modeler}(modeler, data, CVSampler, costFunction, times = 1)
}
\arguments{
\item{modeler}{\code{\link{modeler}} object with the untrained model.}

\item{data}{Datatable with the dataset to use during cross-validation.}

\item{CVSampler}{A cross-validation type \code{\link{sampler}}.}

\item{costFunction}{A costFunction object from \code{\link{costFunction}}}

\item{times}{Amount of times to sample (1 by default)}
}
\value{
The reduced cost function computed for the cross-validation.
}
\description{
Computes the cross validation from a modeler.
Given a cross-validation sample it trains an independent model for each of the
k-folds. It then computes its corresponding cost function and stores it in a list.
The list is then used to compute the reduced cost and therefore returns a simple scalar
instead of a list.
}
\examples{
#Generate data
data.dt <- generateRegressionData(n_samples = 10000, generate_random_weights = FALSE)

#Get relevant variables from data
data.keys <- 1:nrow(data.dt)
target <- "regression_target"# Regression
regressors <- setdiff(names(data.dt), target)

#Define modeler
params <- list(
  "learning_rate" = 0.07,
  "nrounds" = 10,
  "objective" = "regression"
)

modeler <- LGBRegressorModeler(params = params, target = target, regressors = regressors)

#Instantiate cost function
cf <- MSECostFunction()

#Instantiate sampler
cv_sampler <- cvSampler(key = data.keys, folds = 3)

set.seed(273)
reduced_cost <- crossValidate(modeler = modeler,
                             data = data.dt,
                             CVSampler = cv_sampler,
                             costFunction = cf)
print(reduced_cost)

}
\seealso{
\code{\link{expandHyperParams}}
}
