% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{cvGroupSampler}
\alias{cvGroupSampler}
\title{Grouped cross-validation sampler}
\usage{
cvGroupSampler(key, data, identifier, folds)
}
\arguments{
\item{key}{Vector of indices of the data entries (rows) to sample from.}

\item{data}{Dataset containing the identifier-group column.}

\item{identifier}{String indicating the name of the group column.}

\item{folds}{Number of k-folds to apply.}
}
\value{
A cvGroupSampler object.
}
\description{
A cross-validation sampler that instead of randomly partitioning the data,
defines its folds according to a grouping variable, where each fold will
exclude a certain amount of groups.
}
\examples{
set.seed(273)
#Define data (10 groups, 10 rows for each group)
id_var <- c(sapply(paste0("Group_", 1:10), function(x){
return(rep(x, 10))
}, USE.NAMES = FALSE))
measure <- rnorm(n = length(id_var), mean = 10, sd = 1)
data.dt <- data.table(measure = measure, id_var = id_var)

#Instantiate sampler
key <- 1:nrow(data.dt)
identifier <- "id_var"
smplr <- cvGroupSampler(key = key,
                        data = data.dt,
                        identifier = identifier,
                        folds = 3)
#As it can be seen, the data partitions contain all rows from the selected
#groups
s1 <- getSample(smplr)
print(s1)
table(data.dt[s1[, 1], id_var])
table(data.dt[s1[, 2], id_var])
table(data.dt[s1[, 3], id_var])

}
\seealso{
\code{\link{sampler}}

\code{\link{cvSampler}}

\code{\link{cvMixedSampler}}

\code{\link{cvTemporalSampler}}
}
