% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{cvMixedSampler}
\alias{cvMixedSampler}
\title{Mixed data cross-validation sampler}
\usage{
cvMixedSampler(key, fixedKey, folds)
}
\arguments{
\item{key}{Vector of indices for all data, including the fixed data keys
and the remaining (non-fixed) data.}

\item{fixedKey}{Vector of indices indicating which samples should always be
included in every sample. Naturally, \code{length(fixedKey) < length(key)}.}

\item{folds}{Number of k-folds to apply.}
}
\value{
A cvMixedSampler object.
}
\description{
A cross-validation sampler designed to always include a fixed set of rows
in its samples, alongside a shuffled mixture of other samples that do not
always belong to each sample.
}
\examples{
set.seed(273)
#Define dataset
data.dt <- generateStratifiedDomainEstimateData(n_samples = 200)

#Define fixed data keys
fixed_keys <- 1:50#These will always be included
all_keys <- 1:nrow(data.dt)

smplr <- cvMixedSampler(key = all_keys,
fixedKey = fixed_keys,
folds = 5)
#As we can see, the fixed keys are always included
print(getSample(smplr))

}
\seealso{
\code{\link{sampler}}

\code{\link{cvSampler}}

\code{\link{cvGroupSampler}}

\code{\link{cvTemporalSampler}}
}
