% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{cvSampler}
\alias{cvSampler}
\title{Cross-Validation sampler}
\usage{
cvSampler(key, folds)
}
\arguments{
\item{key}{Vector of indices of the data entries (rows) to sample from.}

\item{folds}{Number of k-folds to apply.}
}
\value{
A cvSampler object.
}
\description{
Implementation of cross-validation sampling without replacement.
This sampler generates a total of k_fold samples, each one of them excluding one of the defined folds.
If the key vector is not a multiple of the folds parameter, some partitions will contain NA.
}
\examples{
set.seed(273)
#Define key vector
data.dt <- generateStratifiedDomainEstimateData(n_samples = 50)
key <- 1:nrow(data.dt)
smplr <- cvSampler(key = key, folds = 5)
s1 <- getSample(smplr)

print(s1)

print(data.dt[s1[,1]])

}
\seealso{
\code{\link{sampler}}

\code{\link{cvGroupSampler}}

\code{\link{cvMixedSampler}}

\code{\link{cvTemporalSampler}}
}
