% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{cvTemporalSampler}
\alias{cvTemporalSampler}
\title{Temporal cross-validation sampler}
\usage{
cvTemporalSampler(key, data, identifier, folds, n_valPeriods, orderPeriods)
}
\arguments{
\item{key}{Vector of indices of the data entries (rows) to sample from.}

\item{data}{Complete dataset (train+validation) containing the identifier-period column.}

\item{identifier}{String indicating the name of the period column.}

\item{folds}{Number of k-folds to apply.}

\item{n_valPeriods}{Total amount of periods in validation set.}

\item{orderPeriods}{Chronologically sorted period vector.}
}
\value{
A cvTemporalSampler object.
}
\description{
This sampler implements an expanding window strategy for defining training
folds for a modeler that requires chronologically sorted data, rather than
randomly split data.

For each fold (j), a data threshold is defined as:

\code{thres <- total_periods - (n_valPeriods * folds) + (n_valPeriods * (j - 1)) + 1}

Splitting the sorted train and validation data accordingly. The user should
note that the total amount of periods (\code{length(orderPeriods)}) must be at
least (\code{n_valPeriods * folds + 1}).
}
\examples{
#Define a dataset with 12 periods (months)
years <- c(2023)
total_periods <- c(sapply(years, function(y){
  months <- 1:12
  months[1:9] <- paste0("0", months[1:9])
  return(paste0("MM", months, y))
}, USE.NAMES = FALSE))
target <- rnorm(n = length(total_periods), mean = 10, sd = 1)
data.dt <- data.table::data.table(measure = target, period = total_periods)
print(data.dt)

smplr <- cvTemporalSampler(key = 1:nrow(data.dt),
                           data = data.dt,
                           identifier = "period",
                           folds = 3,
                           n_valPeriods = 2,#Require 2 months for validation
                           orderPeriods = total_periods)

getSample(smplr)

}
\seealso{
\code{\link{sampler}}

\code{\link{cvSampler}}

\code{\link{cvMixedSampler}}

\code{\link{cvGroupSampler}}
}
