% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorModeler.R
\name{directRegressorModeler}
\alias{directRegressorModeler}
\alias{directRegressorModeler.character}
\title{Class constructor for directRegressorModeler.}
\usage{
directRegressorModeler(protoTarget, scaleFactor, target)

\method{directRegressorModeler}{character}(protoTarget, scaleFactor = 1, target)
}
\arguments{
\item{protoTarget}{First estimation of the target
(this values are assumed to be close enough to the target).}

\item{scaleFactor}{Factor to scale the protoTarget by. Set to 1 by default.}

\item{target}{Name of the target variable.}
}
\value{
A directRegressorModeler.
}
\description{
A direct imputation modeler that allows to scale its
predictions by a specified scale factor. By definition, it does not need to
be trained, as the protoTarget and scaleFactor are enough for predicting.
}
\examples{
#Generate data and select only the regression target
set.seed(273)
data.dt <- generateRegressionData(n_samples = 100, generate_random_weights = FALSE)
data.dt <- data.dt[, c("regression_target"), with = FALSE]

#Define a simple protoTarget such as the median of a sample
smlp <- sample(x = 1:nrow(data.dt), size = 50, replace = FALSE)
med <- data.dt[smlp, median(regression_target)]
data.dt[, proto_target := med]
head(data.dt)

modeler <- directRegressorModeler(protoTarget = "proto_target",
                                  scaleFactor = 1,
                                  target = "regression_target")

cf <- MSECostFunction()
evaluate(modeler, data = data.dt, cost_function = cf)

}
\seealso{
\code{\link{regressorModeler}}
}
