% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{estimator}
\alias{estimator}
\title{A collection of estimator functions.}
\description{
Each estimator outputs the estimate and variance for a given target variable.
}
\section{Current implementations}{


\itemize{
\item \code{\link{HTEstimate}}
\item \code{\link{HTDomainEstimate}}
\item \code{\link{REstimate}}
\item \code{\link{RDomainEstimate}}
\item \code{\link{MCSubRBEstimate}}
\item \code{\link{MCSubRBDomainEstimate}}
\item \code{\link{biasVarianceEstimate}}
\item \code{\link{predictionEstimate}}
\item \code{\link{predictionDomainEstimate}}
\item \code{\link{preTrainedEstimate}}
}
}

