% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate.modeler}
\alias{evaluate.prePredModeler}
\alias{evaluate.classifRegModeler}
\alias{evaluate.preproModeler}
\alias{evaluate.memoryModeler}
\title{Built-in modeler evaluation.}
\usage{
evaluate(object, data, cost_function = NULL, subSet = NULL, csample = NULL)

\method{evaluate}{modeler}(object, data, cost_function = NULL, subSet = NULL, csample = NULL)

\method{evaluate}{prePredModeler}(object, data, cost_function = NULL, subSet = NULL, csample = NULL)

\method{evaluate}{classifRegModeler}(object, data, cost_function = NULL, subSet = NULL, csample = NULL)

\method{evaluate}{preproModeler}(object, data, cost_function = NULL, subSet = NULL, csample = NULL)

\method{evaluate}{memoryModeler}(object, data, cost_function = NULL, subSet = NULL, csample = NULL)
}
\arguments{
\item{object}{Modeler  object to predict with.}

\item{data}{Dataset to do the prediction with.}

\item{cost_function}{A costFunction object or a named list of cost functions,
for evaluating complex modelers, such as prePredModeler and classifRegModeler.}

\item{subSet}{List of indices to subset form the dataframe.}

\item{cSample}{An optional argument for compatibility in some specific cases.
Generally, the user should use subSet instead of cSample to define the data subset to evaluate.}
}
\value{
Performance metric for a given modeler, dataset and cost function.
}
\description{
Evaluates a trained modeler object (either a \code{\link{regressorModeler}} or a \code{\link{classifierModeler}}) over a provided dataset and cost function.
}
\examples{
# Generate data
data.dt <- generateRegressionData(n_samples = 1000, generate_random_weights = FALSE)

#Define train and test partitions
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))

#Split train/test ####
sampler <- srsSampler(data.keys, ssize = train_size)#Simple random sampler

train.keys <- getSample(sampler)
test.keys <- setdiff(data.keys, train.keys)

train.dt <- data.dt[train.keys]
test.dt <- data.dt[test.keys]

#Define modeler parameters and modeler itself
target <- "regression_target"# Regression
regressors <- setdiff(names(data.dt), target)

params <- list(
  "learning_rate" = 0.07,
  "nrounds" = 10,
  "objective" = "regression"
)

regressor <- LGBRegressorModeler(params = params, target = target, regressors = regressors)
regressor <- train(regressor, data = train.dt)

#Evaluate R^2 cost function over train and test
cf <- r2CostFunction()
evaluate(object = regressor, data = train.dt, cost_function = cf)
evaluate(object = regressor, data = test.dt, cost_function = cf)

}
