% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expandHyperParams}
\alias{expandHyperParams}
\title{Expand hyper parameters}
\usage{
expandHyperParams(hyperParamsSet)
}
\arguments{
\item{hyperParamSet}{Named list of hyperparameters.}
}
\value{
List of lists. Where each sublist is a possible combination of hyperparameters.
}
\description{
A function to expand a list of hyperparameters into all its possible combinations, generally useful for cross-validation.
}
\examples{
hyperparams <- list(
family = "gaussian",
lambda = 10**(-3:3),
alpha = 0:10 / 10)

combinations <- expandHyperParams(hyperparams)

print(combinations)

}
\seealso{
\code{\link{crossValidate}}

\code{\link{hyperParamCVH2O}}
}
