% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{fullSampler}
\alias{fullSampler}
\title{Full sampler}
\usage{
fullSampler(key, ssize = 1)
}
\arguments{
\item{key}{Vector of indices of the data entries (rows) to sample from.}

\item{ssize}{Integer indicating whether to return all possible combinations
of indices (2) or simply the total original indices (ssize != 2).}
}
\value{
A fullSampler object.
}
\description{
A toy sampler that returns the full set of keys or all possible combinations
(by pairs) of such keys, as long as the ssize parameter is set to 2.
}
\examples{
key <- 1:6

smplr1 <- fullSampler(key = key, ssize = 2)

s1 <- getSample(smplr1)
print(s1)

smplr2 <- fullSampler(key = key)

s2 <- getSample(smplr2)
print(s2)

}
\seealso{
\code{\link{sampler}}

\code{\link{srsSampler}}

\code{\link{ssSampler}}

\code{\link{cvSampler}}

\code{\link{bootstrapSampler}}
}
