% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDatasets.R
\name{generateEstimateData}
\alias{generateEstimateData}
\title{Generate Estimate Data}
\usage{
generateEstimateData(n_samples = 100, generate_random_weights = FALSE)
}
\arguments{
\item{n_samples}{Amount of samples (rows) to return.}

\item{generate_random_weights}{Boolean indicating whether to add a randomised weights column to the dataset.}
}
\value{
data.table with of dim (n_samples, 3) containing three numeric variables, explained above.
}
\description{
A custom function to generate a simple dataset for estimation tasks.
The output contains three columns named Target_1, Target_2 and Target_3, where:

\itemize{
  \item Target_1 follows a Normal(10, 2) distribution
  \item Target_2 follows a Normal(-20, 10) distribution
  \item Target_3 follows a Normal(0, 5) distribution
}
}
\examples{
data <- generateEstimateData(n_samples = 100)
print(data)

}
\seealso{
\code{\link{generateRegressionData}} for simple regression case.

\code{\link{generateStratifiedEstimateData}} for stratified estimation case.

\code{\link{generateStratifiedDomainEstimateData}} for stratified estimation with domains case.

\code{\link{generateModelAssistedRegressionData}} for model-assisted regression (and/or estimation) case.

\code{\link{generateStratifiedDomainModelAssistedEstimateData}} for model-assisted regression (and/or estimation) in a domain-stratified case.
}
