% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDatasets.R
\name{generateModelAssistedRegressionData}
\alias{generateModelAssistedRegressionData}
\title{Generate Model Assisted Regression Data}
\usage{
generateModelAssistedRegressionData(
  n_samples = 100,
  generate_random_weights = FALSE
)
}
\arguments{
\item{n_samples}{Amount of samples (rows) to return.}

\item{generate_random_weights}{Boolean indicating whether to add a randomised weights column to the dataset.}
}
\value{
data.table with dimensions (n_samples, 10) or (n_samples, 11) if generate_random_weights was set to true.
}
\description{
A custom function to generate a simple dataset for testing model-assisted estimation in the MCSubRBEstimator.
 It generates a dataset with two numeric target variables linearly dependent on the remaining columns.
 Some random noise was added to complicate the regression task.
}
\examples{
data <- generateModelAssistedRegressionData(n_samples = 100, generate_random_weights = TRUE)
print(data)

}
\seealso{
\code{\link{generateRegressionData}} for simple regression case.

\code{\link{generateEstimateData}} for estimation case.

\code{\link{generateStratifiedEstimateData}} for stratified estimation case.

\code{\link{generateStratifiedDomainEstimateData}} for stratified estimation with domains case.

\code{\link{generateStratifiedDomainModelAssistedEstimateData}} for model-assisted regression (and/or estimation) in a domain-stratified case.
}
