% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDatasets.R
\name{generateRegressionData}
\alias{generateRegressionData}
\title{Generate Regression Data}
\usage{
generateRegressionData(n_samples = 100, generate_random_weights = TRUE)
}
\arguments{
\item{n_samples}{Amount of samples (rows) to return.}

\item{generate_random_weights}{Boolean indicating whether to add a randomised weights column to the dataset.}
}
\value{
data.table with of dim (n_samples, 12) containing five categorical variables, five numeric variables, a randomized weights column (if generate_random_weights was set to true) and a categorical target variable named regression_target.
}
\description{
A custom function to generate a simple dataset for regression tasks.
}
\examples{
data <- generateRegressionData(n_samples = 100, generate_random_weights = TRUE)
print(data)

}
\seealso{
\code{\link{generateMulticlassClassificationData}} for classification (multiclass) case.

\code{\link{generateBinaryClassificationData}} for classification (binary) case.

\code{\link{generateEstimateData}} for estimation case.

\code{\link{generateModelAssistedRegressionData}} for model-assisted regression (and/or estimation) case.
}
