% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDatasets.R
\name{generateStratifiedDomainEstimateData}
\alias{generateStratifiedDomainEstimateData}
\title{Generate Stratified Domain Estimate Data}
\usage{
generateStratifiedDomainEstimateData(
  n_samples = 100,
  generate_random_weights = FALSE
)
}
\arguments{
\item{n_samples}{Amount of samples (rows) to return.}

\item{generate_random_weights}{Boolean indicating whether to add a randomised weights column to the dataset.}
}
\value{
data.table with dimensions (n_samples, 4) containing two numeric variables and two categorical variables, explained above.
}
\description{
A custom function to generate a simple dataset for estimation tasks.
 The output contains four columns named income, cost_of_living, strata_class and domain_class where:

\itemize{
  \item income follows a lognormal(8, 0.5) distribution, multiplied by strata_class_factor
  \item cost_of_living follows a normal(1000, 200) distribution, multiplied by domain_class_factor
  \item strata_class : Categorical variable with two levels; (Urban, Rural)
  \item domain_class : Categorical variable with three levels; (Province_1, Province_2, Province_3)
}

Where its class factors are:
\itemize{
  \item strata_class_factor = {'Rural': 1, 'Urban': 4}
  \item domain_class_factor = {'Province_1': 2.4, 'Province_2': 1.35, 'Province_3': 0.6}
}
Province_1, Province_2 and Province_3 contain 50%, 30% and 20% of the population respectively. With the following Urban/Rural ratio:
\itemize{
 \item Province_1: 80% Urban, 20% Rural.
 \item Province_2: 30% Urban, 70% Rural.
 \item Province_3: 0% Urban, 100% Rural.
}
}
\examples{
data <- generateStratifiedDomainEstimateData(n_samples = 100)
print(data)

}
\seealso{
\code{\link{generateRegressionData}} for simple regression case.

\code{\link{generateEstimateData}} for estimation case.

\code{\link{generateStratifiedEstimateData}} for stratified estimation case.

\code{\link{generateModelAssistedRegressionData}} for model-assisted regression (and/or estimation) case.

\code{\link{generateStratifiedDomainModelAssistedEstimateData}} for model-assisted regression (and/or estimation) in a domain-stratified case.
}
