% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDatasets.R
\name{generateStratifiedDomainModelAssistedEstimateData}
\alias{generateStratifiedDomainModelAssistedEstimateData}
\title{Generate Stratified Domain Model Assisted Estimate Data}
\usage{
generateStratifiedDomainModelAssistedEstimateData(
  n_samples = 100,
  generate_random_weights = FALSE
)
}
\arguments{
\item{n_samples}{Amount of samples (rows) to return.}
}
\value{
data.table with dimensions (n_samples, 5), explained above and in the generateStratifiedDomainEstimateData documentation.
}
\description{
A custom function to generate a simple dataset for testing model-assisted estimation in a domain-based setting.
It makes use of the generateStratifiedDomainEstimateData function, adding a noisy target variable to the dataset.
}
\examples{
data <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 100)
print(data)

}
\seealso{
\code{\link{generateRegressionData}} for simple regression case.

\code{\link{generateEstimateData}} for estimation case.

\code{\link{generateStratifiedEstimateData}} for stratified estimation case.

\code{\link{generateStratifiedDomainEstimateData}} for stratified estimation with domains case.

\code{\link{generateModelAssistedRegressionData}} for model-assisted regression (and/or estimation) case.
}
