% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDatasets.R
\name{generateStratifiedEstimateData}
\alias{generateStratifiedEstimateData}
\title{Generate Stratified Estimate Data}
\usage{
generateStratifiedEstimateData(
  n_samples = 100,
  generate_random_weights = FALSE
)
}
\arguments{
\item{n_samples}{Amount of samples (rows) to return.}

\item{generate_random_weights}{Boolean indicating whether to add a randomised weights column to the dataset.}
}
\value{
data.table with dimensions (n_samples, 4) containing three numeric variables
        (Target_1, Target_2, Target_3) and one factor variable (strata_class).
}
\description{
A custom function to generate a simple dataset for estimation tasks.
The output contains four columns named Target_1, Target_2, Target_3 and strata_class, where:

\strong{If strata_class == Class_1:}
\itemize{
  \item Target_1 follows a Normal(10, 2) distribution
  \item Target_2 follows a Normal(-20, 10) distribution
  \item Target_3 follows a Normal(0, 5) distribution
}

\strong{If strata_class == Class_2:}
\itemize{
  \item Target_1 follows a Normal(1, 2) distribution
  \item Target_2 follows a Normal(-20, 1) distribution
  \item Target_3 follows a Normal(30, 20) distribution
}

\strong{If strata_class == Class_3:}
\itemize{
  \item Target_1 follows a Normal(20, 2) distribution
  \item Target_2 follows a Normal(-20, 20) distribution
  \item Target_3 follows a Normal(10, 10) distribution
}
\strong{Strata characteristics:}
\itemize{
 \item Target_1 has a different mean and identical standard deviation in each strata.
 \item Target_2 has an identical mean but different standard deviation in each strata.
 \item Target_3 has a different mean and standard deviation in each strata.
}
\strong{The distribution of strata is as follows:}
\itemize{
  \item Class_1 contains around 50\% of the total samples
  \item Class_2 contains around 30\% of the total samples
  \item Class_3 contains around 20\% of the total samples
}
}
\examples{
data <- generateStratifiedEstimateData(n_samples = 100)
print(data)

}
\seealso{
\code{\link{generateRegressionData}} for simple regression case.

\code{\link{generateEstimateData}} for estimation case.

\code{\link{generateStratifiedDomainEstimateData}} for stratified estimation with domains case.

\code{\link{generateModelAssistedRegressionData}} for model-assisted regression (and/or estimation) case.

\code{\link{generateStratifiedDomainModelAssistedEstimateData}} for model-assisted regression (and/or estimation) in a domain-stratified case.
}
