% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClasses.R
\name{getClasses}
\alias{getClasses}
\alias{getClasses.default}
\alias{getClasses.classifierModeler}
\alias{getClasses.preproClassifierModeler}
\title{Get class predictions with a classifierModeler object}
\usage{
getClasses(object, data, class_mapping = maxScoreClass(), subSet = NULL)

\method{getClasses}{default}(object, data, class_mapping = maxScoreClass(), subSet = NULL)

\method{getClasses}{classifierModeler}(object, data, class_mapping = maxScoreClass(), subSet = NULL)

\method{getClasses}{preproClassifierModeler}(object, data, class_mapping = maxScoreClass(), subSet = NULL)
}
\arguments{
\item{object}{\code{\link{classifierModeler}} object to predict with.}

\item{data}{Dataset to do the prediction with.}

\item{class_mapping}{A \code{\link{classMapping}} object, indicating how to map scores to classes}

\item{subSet}{List of indices to subset form the dataframe.}
}
\value{
A vector of class predictions
}
\description{
Computes class predictions from a trained \code{\link{classifierModeler}} object and a frame of (new)
regressors.
}
\examples{
#Define data and model, then train it.
data.dt <- generateBinaryClassificationData(n_samples = 10000, generate_random_weights = FALSE)
target <- "classification_target"
regressors <- setdiff(names(data.dt), target)
params <- list(
  "learning_rate" = 0.07,
  "nrounds" = 10,
  "objective" = "multiclass",#Works better this way with RMLutils
  "num_class" = length(unique(data.dt[, classification_target]))
)

classifier <- LGBClassifierModeler(params = params, target = target, regressors = regressors)
classifier <- train(classifier, data = data.dt)

#Define max score class mapping
cls_mapping_1 <- maxScoreClass(ties.method = "first")

#Get classes given a class mapping
pred_classes1 <- getClasses(object = classifier, data = data.dt, class_mapping = cls_mapping_1)
head(pred_classes1)

#Define binary threshold class mapping
cls_mapping_2 <- binaryThresholdClass(threshold = 0.5, positive_class = "Target_1")

#Get classes given a class mapping
pred_classes2 <- getClasses(object = classifier, data = data.dt, class_mapping = cls_mapping_2)
head(pred_classes2)

}
\seealso{
\code{\link{classMapping}}

\code{\link{maxScoreClass}}

\code{\link{binaryThresholdClass}}

\code{\link{getClassMapping}}
}
