% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClassMapping.R
\name{getClassMapping}
\alias{getClassMapping}
\alias{getClassMapping.maxScoreClass}
\alias{getClassMapping.binaryThresholdClass}
\title{Get class mapping method}
\usage{
getClassMapping(object, scores)

\method{getClassMapping}{maxScoreClass}(object, scores)

\method{getClassMapping}{binaryThresholdClass}(object, scores)
}
\arguments{
\item{object}{A \code{\link{classMapping}} object}

\item{scores}{A named scores matrix, like the ones outputted via \code{\link{getPredictions}} by a \code{\link{classifierModeler}}.}
}
\value{
A vector or list of indices, indicating the classes that have been selected for each sample.
}
\description{
Defines different ways of assigning a class (or group of classes) when given an array of scores/probabilities, outputted from a modeler object.
The user should take into account that while this method may be used to map prediction scores to classes, it is a better practice to use a higher level function such as \code{\link{getClasses}}.
}
\examples{
#Define data and model, then train it.
data.dt <- generateBinaryClassificationData(n_samples = 10000, generate_random_weights = FALSE)
target <- "classification_target"
regressors <- setdiff(names(data.dt), target)
params <- list(
  "learning_rate" = 0.07,
  "nrounds" = 10,
  "objective" = "multiclass",#Works better this way with RMLutils
  "num_class" = length(unique(data.dt[, classification_target]))
)

classifier <- LGBClassifierModeler(params = params, target = target, regressors = regressors)
classifier <- train(classifier, data = data.dt)

#Get prediction scores
pred_scores <- getPredictions(classifier, data)
head(pred_scores)

#Define max score class mapping
cls_mapping <- maxScoreClass(ties.method = "first")

#Get class indices given a class mapping
output <- getClassMapping(cls_mapping, pred_scores)
head(output)#As it can be seen, only the max score classes are selected

#Define binary threshold class mapping
cls_mapping <- binaryThresholdClass(threshold = 0.78, positive_class = "Target_1")

#Get class indices given a class mapping
output <- getClassMapping(cls_mapping, pred_scores)
#Since we demand for a very high threshold in order for Target_1 to be selected,
#in this case all mapped classes correspond to Target_2
head(output)

}
\seealso{
\code{\link{classMapping}}

\code{\link{maxScoreClass}}

\code{\link{binaryThresholdClass}}

\code{\link{getClasses}}
}
