% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCost.R
\name{getCost}
\alias{getCost}
\alias{getCost.r2CostFunction}
\alias{getCost.squareBiasCostFunction}
\alias{getCost.MSECostFunction}
\alias{getCost.groupTotalErrorCostFunction}
\alias{getCost.groupMSECostFunction}
\alias{getCost.MCSubRBMSECostFunction}
\alias{getCost.holdoutTemporalCostFunction}
\alias{getCost.confusionMatrixCostFunction}
\alias{getCost.accuracyCostFunction}
\alias{getCost.precisionCostFunction}
\alias{getCost.recallCostFunction}
\alias{getCost.f1ScoreCostFunction}
\alias{getCost.ROCAUCCostFunction}
\title{Evaluates cost function}
\usage{
getCost(object, data, modeler, fSample, cSample)

\method{getCost}{r2CostFunction}(object, data, modeler, fSample, cSample)

\method{getCost}{squareBiasCostFunction}(object, data, modeler, fSample, cSample)

\method{getCost}{MSECostFunction}(object, data, modeler, fSample, cSample)

\method{getCost}{groupTotalErrorCostFunction}(object, data, modeler, fSample, cSample)

\method{getCost}{groupMSECostFunction}(object, data, modeler, fSample, cSample)

\method{getCost}{groupTotalErrorCostFunction}(object, data, modeler, fSample, cSample)

\method{getCost}{MCSubRBMSECostFunction}(object, data, modeler, fSample, cSample)

\method{getCost}{holdoutTemporalCostFunction}(object, data, modeler, fSample, cSample)

\method{getCost}{confusionMatrixCostFunction}(object, data, modeler, fSample, cSample)

\method{getCost}{accuracyCostFunction}(object, data, modeler, fSample, cSample)

\method{getCost}{precisionCostFunction}(object, data, modeler, fSample, cSample)

\method{getCost}{recallCostFunction}(object, data, modeler, fSample, cSample)

\method{getCost}{f1ScoreCostFunction}(object, data, modeler, fSample, cSample)

\method{getCost}{ROCAUCCostFunction}(object, data, modeler, fSample, cSample)
}
\arguments{
\item{object}{\code{\link{costFunction}} object.}

\item{data}{Datatable containing dataset to work with.}

\item{modeler}{Model object used whose performance is to be studied.}

\item{fSample}{All indices of entries obtained by \code{\link{sampler}}.}

\item{cSample}{Complementary indices to the ones obtained by sampler.}
}
\value{
A named list containing cost and additional metadata, useful for getReducedCost.
}
\description{
Returns the evaluation of the cost function in the given complementary
sample.
}
\examples{
# Regression example ####
#Generate data
data.dt <- generateRegressionData(n_samples = 1000, generate_random_weights = FALSE)

#Instantiate cost function
cf <- r2CostFunction(weightsColumn = "")

target <- "regression_target"# Regression

regressors <- setdiff(names(data.dt), target)

#Define modeler
params <- list(
  "learning_rate" = 0.07,
  "nrounds" = 10,
  "objective" = "regression"
)

regressor <- LGBRegressorModeler(params = params, target = target, regressors = regressors)
regressor <- train(regressor, data = data.dt)

cost <- getCost(object = cf,
                data = data.dt,
                modeler = regressor,
                fSample = 1:nrow(data.dt),
                cSample = 1:nrow(data.dt))
print(cost)

# Classification example ####
#Generate data
data.dt <- generateMulticlassClassificationData(n_samples = 1000)
#Instantiate cost function
cf <- confusionMatrixCostFunction(weightsColumn = "",
                                  class_map = maxScoreClass(),
                                  normalize = NULL,
                                  class_names = levels(data.dt[["classification_target"]]))

#Train simple modeler
target <- "classification_target"
regressors <- setdiff(names(data.dt), target)
params <- list(
  "learning_rate" = 0.07,
  "nrounds" = 10,
  "objective" = "multiclass",
  "num_class" = length(unique(data.dt[, classification_target]))
)

classifier <- LGBClassifierModeler(params = params, target = target, regressors = regressors)
classifier <- train(classifier, data = data.dt)

#Evaluate with getCost
cost <- getCost(object = cf, data = data.dt, modeler = classifier, cSample = 1:nrow(data.dt))
print(cost)

}
\seealso{
\code{\link{costFunction}}

\code{\link{getReducedCost}}

\code{\link{r2CostFunction}}

\code{\link{confusionMatrixCostFunction}}
}
