% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredictions.R
\name{getPredictions}
\alias{getPredictions}
\alias{getPredictions.directModeler}
\alias{getPredictions.HTModeler}
\alias{getPredictions.strataHTModeler}
\alias{getPredictions.strataMeanModeler}
\alias{getPredictions.H2OModeler}
\alias{getPredictions.benchModeler}
\alias{getPredictions.preproModeler}
\alias{getPredictions.LGBModeler}
\alias{getPredictions.rangerModeler}
\alias{getPredictions.catboostModeler}
\alias{getPredictions.miceModeler}
\alias{getPredictions.classifRegModeler}
\alias{getPredictions.prePredModeler}
\alias{getPredictions.memoryModeler}
\title{Predicts with a modeler object}
\usage{
getPredictions(object, data, subSet = NULL)

\method{getPredictions}{directModeler}(object, data, subSet = NULL)

\method{getPredictions}{HTModeler}(object, data, subSet = NULL)

\method{getPredictions}{strataHTModeler}(object, data, subSet = NULL)

\method{getPredictions}{strataMeanModeler}(object, data, subSet = NULL)

\method{getPredictions}{H2OModeler}(object, data, subSet = NULL)

\method{getPredictions}{benchModeler}(object, data, subSet = NULL)

\method{getPredictions}{preproModeler}(object, data, subSet = NULL)

\method{getPredictions}{LGBModeler}(object, data, subSet = NULL)

\method{getPredictions}{rangerModeler}(object, data, subSet = NULL)

\method{getPredictions}{catboostModeler}(object, data, subSet = NULL)

\method{getPredictions}{miceModeler}(object, data = NULL, subSet = NULL)

\method{getPredictions}{classifRegModeler}(object, data, subSet = NULL)

\method{getPredictions}{prePredModeler}(object, data, subSet = NULL)

\method{getPredictions}{memoryModeler}(object, data, subSet = NULL)
}
\arguments{
\item{object}{Model object to predict with.}

\item{data}{Dataset to do the prediction with.}

\item{subSet}{List of indices to subset form the dataframe.}
}
\value{
A vector of predictions
}
\description{
Computes predictions from a trained \code{\link{modeler}} object and a frame of (new)
regressors.
}
\examples{
#Generate data
data.dt <- generateRegressionData(n_samples = 1000, generate_random_weights = FALSE)

#Define train/test partitions
set.seed(273)
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

#Define modeler
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)
target <- "regression_target"
regressors <- setdiff(colnames(data.dt), target)

modeler <- rangerRegressorModeler(params = params, target = target, regressors = regressors)

#Train, predict and evaluate
modeler <- train(object = modeler, data = data.dt, subSet = train.keys)

preds <- getPredictions(modeler, data = data.dt, subSet = train.keys)

}
