% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPreproData}
\alias{getPreproData}
\alias{getPreproData.preproObject}
\title{Method to apply preprocessing to data, given a preproObject}
\usage{
getPreproData(object, data)

\method{getPreproData}{preproObject}(object, data)
}
\arguments{
\item{object}{preproObject to be used. It should contain a preprocessing function and a set of parameters.}

\item{data}{dataset to be preprocessed.}
}
\value{
preproObject with preprocessed data.
}
\description{
Applies preprocessing to the given data, saving the data inside the object. Additionally, saves the new regressors in case some were added during the preprocessing.
}
\examples{
#Generate data
data.dt <- generateRegressionData(n_samples = 1000,
                                  generate_random_weights = TRUE)
head(data.dt)

#Define relevant variables
target <- "regression_target"
weightsColumn <- "weights"
regressors <- setdiff(colnames(data.dt), c(target, weightsColumn))

modeler <- list("target" = target,
                "regressors" = regressors,
                "weightsColumn" = weightsColumn)

#Define preproObject
#Preprocessing parameters
prepro_params <- list(modeler_type = NULL, pretrained_min_max = NULL)

#Define preproObject
prepro_object <- preproObject(preproFunction = minMaxScalerPreproFunction,
                              params = prepro_params)

#Assign relevant metadata to preproObject
prepro_object <- assignExternalAttributes(modeler = modeler,
                                          object = prepro_object)

#Call getPreproData to obtain preproObject with preprocessed data
prepro_object <- getPreproData(object = prepro_object, data = data.dt)

#data is saved within prepro_object
head(prepro_object$data)

}
\seealso{
\code{\link{preproRegressorModeler}}

\code{\link{preproClassifierModeler}}

\code{\link{assignExternalAttributes}}

\code{\link{preproObject}}
}
