% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReducedCost.R
\name{getReducedCost}
\alias{getReducedCost}
\alias{getReducedCost.r2CostFunction}
\alias{getReducedCost.squareBiasCostFunction}
\alias{getReducedCost.domainMSECostFunction}
\alias{getReducedCost.MSECostFunction}
\alias{getReducedCost.groupMSECostFunction}
\alias{getReducedCost.groupTotalErrorCostFunction}
\alias{getReducedCost.MCSubRBMSECostFunction}
\alias{getReducedCost.holdoutTemporalCostFunction}
\alias{getReducedCost.confusionMatrixCostFunction}
\alias{getReducedCost.accuracyCostFunction}
\alias{getReducedCost.precisionCostFunction}
\alias{getReducedCost.recallCostFunction}
\alias{getReducedCost.f1ScoreCostFunction}
\alias{getReducedCost.ROCAUCCostFunction}
\title{Reduction operator for cost function classes.}
\usage{
getReducedCost(object, data)

\method{getReducedCost}{r2CostFunction}(object, data)

\method{getReducedCost}{squareBiasCostFunction}(object, data)

\method{getReducedCost}{domainMSECostFunction}(object, data)

\method{getReducedCost}{MSECostFunction}(object, data)

\method{getReducedCost}{groupMSECostFunction}(object, data)

\method{getReducedCost}{groupTotalErrorCostFunction}(object, data)

\method{getReducedCost}{MCSubRBMSECostFunction}(object, data)

\method{getReducedCost}{holdoutTemporalCostFunction}(object, data)

\method{getReducedCost}{confusionMatrixCostFunction}(object, data)

\method{getReducedCost}{accuracyCostFunction}(object, data)

\method{getReducedCost}{precisionCostFunction}(object, data)

\method{getReducedCost}{recallCostFunction}(object, data)

\method{getReducedCost}{f1ScoreCostFunction}(object, data)

\method{getReducedCost}{ROCAUCCostFunction}(object, data)

\method{getReducedCost}{groupTotalErrorCostFunction}(object, data)
}
\arguments{
\item{object}{\code{\link{costFunction}} object.}

\item{data}{List of cross-validation cost results.}
}
\description{
Summarizes the cost over the different cross validation subsets.
}
\examples{
#Usage outside crossValidate:
#Generate data
data.dt <- generateRegressionData(n_samples = 1000, generate_random_weights = FALSE)
#Get relevant variables from data
data.keys <- 1:nrow(data.dt)
target <- "regression_target"# Regression
regressors <- setdiff(names(data.dt), target)
#Define modeler
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)
target <- "regression_target"
regressors <- setdiff(colnames(data.dt), target)

modeler <- rangerRegressorModeler(params = params, target = target, regressors = regressors)
#Instantiate cost function
cf <- MSECostFunction()
#Instantiate sampler
set.seed(273)
cv_sampler <- cvSampler(key = data.keys, folds = 3)
#Define folds to train the modeler
data_folds <- getSample(cv_sampler)

#Train modeler with different folds and save cost for each
cost_list <- list()
for (i in 1:dim(data_folds)[2]){
  iSample <- data_folds[,i]
  iSample <- sort(iSample[!is.na(iSample)])
  cSample <- sort(setdiff(cv_sampler$key, iSample))
  modeler <- train(modeler, data.dt, iSample)
  cost_list[[i]] <- getCost(cf, data.dt, modeler,
                            cv_sampler$key, cSample)
  modeler <- modelClean(modeler)
}

#Obtain averaged cost
getReducedCost(cf, cost_list)


#Usage with crossValidate
set.seed(273)
crossValidate(modeler = modeler,
              data = data.dt,
              CVSampler = cv_sampler,
              costFunction = cf)

}
\seealso{
\code{\link{costFunction}}

\code{\link{getCost}}
}
