% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegressors.R
\name{getRegressors}
\alias{getRegressors}
\alias{getRegressors.default}
\alias{getRegressors.preproModeler}
\alias{getRegressors.classifRegModeler}
\alias{getRegressors.benchModeler}
\alias{getRegressors.prePredModeler}
\alias{getRegressors.memoryModeler}
\title{Returns regressors}
\usage{
getRegressors(object)

\method{getRegressors}{default}(object)

\method{getRegressors}{preproModeler}(object)

\method{getRegressors}{classifRegModeler}(object)

\method{getRegressors}{benchModeler}(object)

\method{getRegressors}{prePredModeler}(object)

\method{getRegressors}{memoryModeler}(object)
}
\arguments{
\item{object}{Modeler object}
}
\value{
A list with the regressors of the \code{\link{regressorModeler}} or \code{\link{classifierModeler}} object
}
\description{
Returns the regressors of a modeler object
}
\examples{
mdl <- H2OModeler("glm", list(lambda=0.0),"d" , c("a", "b", "c"))
getRegressors(mdl)

}
