% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSample.R
\name{getSample}
\alias{getSample}
\alias{getSample.srsSampler}
\alias{getSample.ssSampler}
\alias{getSample.bootstrapSampler}
\alias{getSample.fullSampler}
\alias{getSample.cvSampler}
\alias{getSample.cvMixedSampler}
\alias{getSample.cvGroupSampler}
\alias{getSample.cvTemporalSampler}
\title{Gets a matrix with samples.}
\usage{
getSample(object, times = 1)

\method{getSample}{srsSampler}(object, times = 1)

\method{getSample}{ssSampler}(object, times = 1)

\method{getSample}{bootstrapSampler}(object, times = 1)

\method{getSample}{fullSampler}(object, times = 1)

\method{getSample}{cvSampler}(object, times = 1)

\method{getSample}{cvMixedSampler}(object, times = 1)

\method{getSample}{cvGroupSampler}(object, times = 1)

\method{getSample}{cvTemporalSampler}(object, times = 1)
}
\arguments{
\item{object}{\code{\link{sampler}} object}

\item{times}{Number of times to sample from data.}
}
\value{
A matrix with the samples. Each column of the data being the n-th
sample taken and the rows the sample results of each sample taken.
 When using stratified sampling, the rows are returned ordered by the strates.
}
\description{
Obtains a sample according the sampling scheme from a sampler object.
}
\examples{
set.seed(273)
#Define simple random sampler and get a sample
sampler1 <- srsSampler(key = 1:15, ssize = 10)
getSample(sampler1, times = 2)

#Define cross-validation sampler and get a sample
sampler2 <- cvSampler(key = 1:25, folds = 5)
getSample(sampler2, times = 1)

}
\seealso{
\code{\link{srsSampler}}

\code{\link{cvSampler}}
}
