% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTarget.R
\name{getTarget}
\alias{getTarget}
\alias{getTarget.benchModeler}
\alias{getTarget.default}
\alias{getTarget.prePredModeler}
\title{Returns target variables}
\usage{
getTarget(object)

\method{getTarget}{benchModeler}(object)

\method{getTarget}{default}(object)

\method{getTarget}{prePredModeler}(object)
}
\arguments{
\item{object}{\code{\link{modeler}} object.}
}
\value{
The target attribute of the modeler object.
}
\description{
Returns a vector of targets variables from a modeler.
Note that modelers with intermediate target variables such as
\code{\link{prePredRegressorModeler}}and
\code{\link{classifRegRegressorModeler}} output their main target variable.

On the other hand, benchRegressorModeler returns a vector of targets (with
each component being the main target of its internal modelers).
}
\examples{
mdl <- H2ORegressorModeler("glm", list(lambda=0.0),"d" , c("a", "b", "c"))
getTarget(mdl)

}
