% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeightsColumn.R
\name{getWeightsColumn}
\alias{getWeightsColumn}
\alias{getWeightsColumn.default}
\alias{getWeightsColumn.preproModeler}
\alias{getWeightsColumn.classifRegModeler}
\alias{getWeightsColumn.prePredRegressorModeler}
\alias{getWeightsColumn.memoryRegressorModeler}
\alias{getWeightsColumn.benchModeler}
\title{Returns weightsColumn}
\usage{
getWeightsColumn(object)

\method{getWeightsColumn}{default}(object)

\method{getWeightsColumn}{preproModeler}(object)

\method{getWeightsColumn}{classifRegModeler}(object)

\method{getWeightsColumn}{prePredRegressorModeler}(object)

\method{getWeightsColumn}{memoryRegressorModeler}(object)

\method{getWeightsColumn}{benchModeler}(object)
}
\arguments{
\item{object}{A modeler object.}
}
\value{
The weightsColumn attribute (or attributes) of the modeler object.
}
\description{
Returns a string or a vector of weightsColumn attributes from a modeler.
Generally, it will return a single string (or a NULL value), but since some
complex modelers contain multiple modelers internally
(which may have a different weights column each), a vector would be returned
in that case.

Note that if some of the multiple weightColumns match, only unique values
would be returned.

Besides other getters such as getTarget and getRegressors, which have a
paired setter method, this one does not have it, because it is not a good
practice to change a weightsColumn attribute after instantiating a modeler.
}
\examples{
modeler <- HTRegressorModeler(target = "income",
                              weightsColumn = "weights")
getWeightsColumn(modeler)

}
