% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costFunctions.R
\name{groupMSECostFunction}
\alias{groupMSECostFunction}
\alias{groupMSECostFunction.character}
\title{Group MSE cost function}
\usage{
groupMSECostFunction(
  weightsColumn,
  groupWeightsFunction,
  groupFunction,
  groupColumn
)

\method{groupMSECostFunction}{character}(
  weightsColumn,
  groupWeightsFunction,
  groupFunction,
  groupColumn
)
}
\arguments{
\item{weightsColumn}{Name of the weights column in the data. "" by default.}

\item{groupWeightsFunction}{Function to average/combine weights for an entire group.}

\item{groupFunction}{Function to average/combine target values for an entire group.}

\item{groupColumn}{Name of the group-defining column in the data, sush as a strata or a domain.}
}
\description{
A cost function that calculates the weighted MSE of a function of each group element.
That is, after predicting, each group is converted into a single metric,
such as a mean or a sum, and the MSE is calculated comparing the average/sum/other transform
of the predictions and original values.

Afterwards, all MSEs are averaged with the data weights (if they exist).
}
\examples{
#Generate data
data.dt <- generateRegressionData(n_samples = 1000)

data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))

#Split train/test #
sampler <- srsSampler(data.keys, ssize = train_size)

train.keys <- getSample(sampler)
test.keys <- setdiff(data.keys, train.keys)

train.dt <- data.dt[train.keys]
test.dt <- data.dt[test.keys]

#Define modeler
target <- "regression_target"
regressors <- setdiff(names(data.dt), target)
params <- list(
  "learning_rate" = 0.07,
  "nrounds" = 10,
  "objective" = "regression"
)

regressor <- LGBRegressorModeler(params = params, target = target, regressors = regressors)
regressor <- train(regressor, data = train.dt)

#Define cost function and get cost

cf <- groupMSECostFunction(weightsColumn = "weights",
                           groupWeightsFunction = sum,
                           groupColumn = "cat_1",
                           groupFunction = mean)

cost <- getCost(object = cf, data = data.dt, modeler = regressor, cSample = test.keys)
print(cost)

}
\seealso{
\code{\link{costFunction}}

\code{\link{getCost}}
}
