% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costFunctions.R
\name{groupTotalErrorCostFunction}
\alias{groupTotalErrorCostFunction}
\alias{groupTotalErrorCostFunction.character}
\title{Group absolute error cost function}
\usage{
groupTotalErrorCostFunction(weightsColumn, groupVars)

\method{groupTotalErrorCostFunction}{character}(weightsColumn, groupVars)
}
\arguments{
\item{weightsColumn}{Name of the weights column in the data. "" by default.}

\item{groupVars}{Vector containing the variables to group.}
}
\description{
A cost function that calculates the weighted absolute error by group, for each
group element, and then combines it to obtain the total absolute prediction error.
}
\examples{
#Generate data
data.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 1000,
                                                             generate_random_weights = TRUE)
data.keys <- 1:nrow(data.dt)
set.seed(273)

data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))

#Split train/test
sampler <- srsSampler(data.keys, ssize = train_size)

train.keys <- getSample(sampler)
test.keys <- setdiff(data.keys, train.keys)

train.dt <- data.dt[train.keys]
test.dt <- data.dt[test.keys]

#Define modeler
target <- "target"
regressors <- setdiff(names(data.dt), target)
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)

modeler <- rangerRegressorModeler(params = params, target = target, regressors = regressors)
modeler <- train(modeler, data = train.dt)

cf <- groupTotalErrorCostFunction(weightsColumn = "weights",
                                  groupVars = c("strata_class", "domain_class"))

cost <- getCost(object = cf,
                data = data.dt,
                modeler = modeler,
                cSample = test.keys)
print(cost)

}
\seealso{
\code{\link{costFunction}}

\code{\link{getCost}}
}
