% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierModeler.R
\name{H2OClassifierModeler}
\alias{H2OClassifierModeler}
\alias{H2OClassifierModeler.character}
\alias{H2OClassifierModeler.data.frame}
\title{Class constructor for H2OClassifierModeler}
\usage{
H2OClassifierModeler(model, params, target, regressors, DF)

\method{H2OClassifierModeler}{character}(model, params, target, regressors)

\method{H2OClassifierModeler}{data.frame}(DF, regressors)
}
\arguments{
\item{model}{Name of H2O model to be used. The following models are available:
\itemize{
  \item adaBoost (Binary classification)
  \item decision_tree (Binary classification)
  \item deeplearning
  \item gam (General additive model)
  \item gbm (Gradient boosting model)
  \item glm (Generalized linear model)
  \item naiveBayes
  \item psvm (Support vector machine, Binary classification)
  \item randomForest
  \item rulefit
  \item xgboost (Only works on linux machines)
}}

\item{params}{List of parameters to be used for training an H2O model.}

\item{target}{String indicating the name of the target variable in the data.}

\item{regressors}{List or vector of strings indicating the regressor column names in the data.}

\item{DF}{Alternate H2O modeler definition.}
}
\value{
An H2OClassifierModeler object.
}
\description{
H2OModeler implementation for classification tasks.
This modeler offers access to the H2O library, but requires the user to run an H2O cluster for it to work.
}
\examples{
#NOTE: Running this tutorial may take a few seconds, since the
#H2O cloud needs to start
#Generate data
data.dt <- generateBinaryClassificationData(n_samples = 1000, generate_random_weights = FALSE)

#Define train/test partitions
set.seed(273)
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

target <- "classification_target"
regressors <- setdiff(colnames(data.dt), target)

#Start H2O cloud (change port if that one is not available)
h2o::h2o.init(port = 54321, startH2O = TRUE)

h2o_model <- "adaBoost"
params <- list()

#Define and train modeler
modeler <- H2OClassifierModeler(model = h2o_model,
                                params = params,
                                target = target,
                                regressors = regressors)

modeler <- train(object = modeler,
                 data = data.dt,
                 subSet = train.keys)

#Predict
#preds <- getPredictions(object = modeler,
#                        data = data.dt,
#                        subSet = train.keys)

#clss <- getClasses(object = modeler,
#                   data = data.dt,
#                   subSet = train.keys,
#                   class_mapping = maxScoreClass())

#Evaluate
cf <- accuracyCostFunction()

getCost(object = cf,
        data = data.dt,
        modeler = modeler,
        fSample = 1:nrow(data.dt),
        cSample = test.keys)

#evaluate(object = modeler,
#         data = data.dt,
#         cost_function = cf,
#         subSet = test.keys)

h2o::h2o.shutdown(prompt = FALSE)


}
\seealso{
\code{\link{classifierModeler}}
}
