% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorModeler.R
\name{H2ORegressorModeler}
\alias{H2ORegressorModeler}
\alias{H2ORegressorModeler.character}
\alias{H2ORegressorModeler.data.frame}
\title{Class constructor for H2ORegressorModeler}
\usage{
H2ORegressorModeler(model, params, target, regressors, DF)

\method{H2ORegressorModeler}{character}(model, params, target, regressors)

\method{H2ORegressorModeler}{data.frame}(DF, regressors)
}
\arguments{
\item{model}{Name of H2O model to be used. The following models are available:
\itemize{
  \item deeplearning
  \item gam (General additive model)
  \item gbm (Gradient boosting model)
  \item glm (Generalized linear model)
  \item randomForest
  \item rulefit
  \item xgboost (Only works on linux machines)
}}

\item{params}{List of parameters to be used for training an H2O model.}

\item{target}{String indicating the name of the target variable in the data.}

\item{regressors}{List or vector of strings indicating the regressor column names in the data.}

\item{DF}{Alternate H2O modeler definition.}
}
\value{
An H2ORegressorModeler object.
}
\description{
H2OModeler implementation for regression tasks.
This modeler offers access to the H2O library, but requires the user to run an H2O cluster for it to work.
}
\examples{
#NOTE: Running this tutorial may take a few seconds, since the
#H2O cloud needs to start
#Generate data
data.dt <- generateRegressionData(n_samples = 1000, generate_random_weights = FALSE)

#Define train/test partitions
set.seed(273)
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

target <- "regression_target"
regressors <- setdiff(colnames(data.dt), target)

#Start H2O cloud (change port if that one is not available)
h2o::h2o.init(port = 54321, startH2O = TRUE)

#Define modeler
h2o_model <- "randomForest"
params <- list()
modeler <- H2ORegressorModeler(model = h2o_model, params = params, target = target, regressors = regressors)

#Train modeler (note that data can either be a data.table or an H2OFrame)
modeler <- train(modeler, data = data.dt, subSet = train.keys)

predictions <- getPredictions(modeler, data = data.dt, subSet = train.keys)

cf <- r2CostFunction()

evaluate(object = modeler, data = data.dt, cost_function = cf, subSet = test.keys)

#Close H2O cloud
h2o::h2o.shutdown(prompt = FALSE)

}
\seealso{
\code{\link{regressorModeler}}

\code{\link{H2OClassifierModeler}}
}
