% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costFunctions.R
\name{holdoutTemporalCostFunction}
\alias{holdoutTemporalCostFunction}
\alias{holdoutTemporalCostFunction.costFunction}
\title{Holdout temporal cost function}
\usage{
holdoutTemporalCostFunction(costFunction, cvTemporalSampler)

\method{holdoutTemporalCostFunction}{costFunction}(costFunction, cvTemporalSampler)
}
\arguments{
\item{costFunction}{A costFunction object.}

\item{cvTemporalSampler}{A cvTemporalSampler object.}
}
\description{
A complex costFunction that takes a simpler costFunction and combines it with
a \code{\link{cvTemporalSampler}}, useful for evaluating data with some sort
of temporal dependence.
}
\examples{
#Define a dataset with 36 periods (months)
years <- c(2023, 2024, 2025)
total_periods <- c(sapply(years, function(y){
  months <- 1:12
  months[1:9] <- paste0("0", months[1:9])
  return(paste0("MM", months, y))
}, USE.NAMES = FALSE))

#Generate data for each period
data.dt <- rbindlist(lapply(total_periods, function(p){
  dt <- generateRegressionData(n_samples = 100,
                               generate_random_weights = FALSE)
  dt[, period := p]
  return(dt)
}))
head(data.dt)

smplr <- cvTemporalSampler(key = 1:nrow(data.dt),
                           data = data.dt,
                           identifier = "period",
                           folds = 5,
                           n_valPeriods = 6,#Require 2 months for validation
                           orderPeriods = total_periods)

#Define modeler and train it
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)
target <- "regression_target"
regressors <- setdiff(colnames(data.dt), target)

modeler <- rangerRegressorModeler(params = params,
                                  target = target,
                                  regressors = regressors)

modeler <- train(modeler, data = data.dt)

#Define holdoutTemporalCostFunction
cf <- holdoutTemporalCostFunction(costFunction = MSECostFunction(),
                                  cvTemporalSampler = smplr)

cost <- getCost(object = cf,
                data = data.dt,
                modeler = modeler,
                cSample = 1:nrow(data.dt))
print(cost$regression_target$cost)

cost <- crossValidate(modeler = modeler,
                      data = data.dt,
                      CVSampler = cvSampler(key = data.keys, folds = 3),
                      costFunction = cf)
print(cost)


}
\seealso{
\code{\link{costFunction}}

\code{\link{getCost}}
}
