% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{HTDomainEstimate}
\alias{HTDomainEstimate}
\alias{HTDomainEstimate.sampler}
\title{Horvitz-Thompson estimator for domains.}
\usage{
HTDomainEstimate(object, sampleKeys, sampleData, domains)

\method{HTDomainEstimate}{sampler}(object, sampleKeys, sampleData, domains)
}
\arguments{
\item{object}{A \code{\link{sampler}} object.}

\item{sampleKeys}{List of indices of the sample.}

\item{sampleData}{Data sampled from the full population.}

\item{domains}{List, matrix or data.table of domains in the sampled data.}
}
\value{
A list with the Horvitz-Thompson estimation of the total and its
variance estimation.
}
\description{
Computes the Horvitz-Thompson estimator for domains according to the
sampling scheme and sample.
}
\examples{
set.seed(273)
#Define data to use
data.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 1000,
                                                             generate_random_weights = FALSE)
head(data.dt)

#Define simple random sampler
key <- 1:nrow(data.dt)
srsmplr <- srsSampler(key = key, ssize = 800)
sample_1 <- getSample(srsmplr)
sample_data.dt <- data.dt[sample_1]
domains <- sample_data.dt[, domain_class]
target_cols <- c("income", "cost_of_living")

#HT estimate for simple random sampling
HTDomainEstimate(object = srsmplr,
                 sampleKeys = sample_1,
                 sampleData = sample_data.dt[, ..target_cols],
                 domains = domains)
print(paste0("Real total value of income: ", std_ma_data.dt[, sum(income)]))
print(paste0("Real total value of  cost_of_living: ", std_ma_data.dt[, sum(cost_of_living)]))

#Use same data with ssSampler
sssmplr <- ssSampler(key = key,
                     strata = data.dt[, strata_class],
                     ssize = c("Rural" = 408, "Urban" = 392))
sample_2 <- getSample(srsmplr)
sample_data_2.dt <- data.dt[sample_2]
domains2 <- sample_data_2.dt[, domain_class]
target_cols <- c("income", "cost_of_living")

#HT domain estimate for stratified random sampling
HTDomainEstimate(object = sssmplr,
                 sampleKeys = sample_2,
                 sampleData = sample_data_2.dt[, ..target_cols],
                 domains = domains2)
print(paste0("Real total value of income: ", std_ma_data.dt[, sum(income)]))
print(paste0("Real total value of  cost_of_living: ", std_ma_data.dt[, sum(cost_of_living)]))

}
\seealso{
\code{\link{estimator}}

\code{\link{HTEstimate}}

\code{\link{RDomainEstimate}}
}
