% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{HTEstimate}
\alias{HTEstimate}
\alias{HTEstimate.srsSampler}
\alias{HTEstimate.ssSampler}
\title{Estimate Horvitz-Thompson estimator.}
\usage{
HTEstimate(object, sampleKeys, sampleData)

\method{HTEstimate}{srsSampler}(object, sampleKeys, sampleData)

\method{HTEstimate}{ssSampler}(object, sampleKeys, sampleData)
}
\arguments{
\item{object}{sampler object to sample from}

\item{sampleKeys}{vector of indices of the sample}

\item{sampleData}{Data sampled from the full population}
}
\value{
A list with the Horvitz-Thompson estimation of the total and its
variance estimation for each variable of the dataframe.
}
\description{
Computes the Horvitz-Thompson estimator according to the sampling scheme
and sample. Returns the estimation for all the variables of the sampled data.
}
\examples{
set.seed(273)
#Generate simple data for estimation purposes
data.dt <- generateEstimateData(n_samples = 1000,
                                generate_random_weights = FALSE)
head(data.dt)

key <- 1:nrow(data.dt)
srsmplr <- srsSampler(key = key, ssize = 800)
sample_1 <- getSample(srsmplr)

#HT estimate for simple random sampling
estim1 <- HTEstimate(object = srsmplr,
                     sampleKeys = sample_1,
                     sampleData = data.dt[sample_1])
print(estim1)

#Theoretical values are known (see help(generateEstimateData) for more details)
c("Target_1" = 10 * 1000,
  "Target_2" = -20 * 1000,
  "Target_3" = 0)

#Standard error of estimate
sqrt(estim1$variance)

#Generate simple stratified data for estimation purposes
data2.dt <- generateStratifiedEstimateData(n_samples = 1000,
                                           generate_random_weights = FALSE)
head(data2.dt)

key = 1:nrow(data2.dt)
sssmplr <- ssSampler(key = key,
                     strata = data2.dt[, strata_class],
                     ssize = c("Class_1" = 400, "Class_2" = 240, "Class_3" = 160))
sample_2 <- getSample(sssmplr)
#Define target columns
target_cols <- c("Target_1", "Target_2", "Target_3")
#HT estimate for stratified random sampling
estim2 <- HTEstimate(object = sssmplr,
                     sampleKeys = sample_2,
                     sampleData = data2.dt[sample_2, ..target_cols])
print(estim2)

#Theoretical values are known (see help(generateStratifiedEstimateData) for more details)
c("Target_1" = 500 * 10 + 300 * 1 + 200 * 20,
  "Target_2" = 500 * (-20) + 300 * (-20) + 200 * (-20),
  "Target_3" = 500 * 0 + 300 * 30 + 200 * 10)

#Standard error of estimate
sqrt(estim2$variance)

}
\seealso{
\code{\link{estimator}}

\code{\link{HTDomainEstimate}}

\code{\link{REstimate}}
}
