% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorModeler.R
\name{HTRegressorModeler}
\alias{HTRegressorModeler}
\alias{HTRegressorModeler.character}
\title{Class constructor for HTRegressorModeler}
\usage{
HTRegressorModeler(weightsColumn, target)

\method{HTRegressorModeler}{character}(weightsColumn, target)
}
\arguments{
\item{weightsColumn}{Name of the weights column.}

\item{target}{Name of the target variable.}
}
\value{
An HTRegressorModeler object.
}
\description{
A regressor modeler that assumes that the target variable is
somewhat proportional to its sampling weight.

A mean HT factor is calculated with the train data:

\deqn{f_{HT} = \sum_{i = 1}^{n}\frac{\left(\frac{1}{\pi_{i}} - 1\right)}{\left(n - \sum_{j = 1}^{n} \pi_{j}\right)} \cdot y_{i}}

Which is then used for predicting, dividing it by the sampling weight of the
units to predict, assuming that weights are defined as
\eqn{w_{i} = \frac{1}{\pi_{i}}}, where \eqn{\pi_{i}} is the
sampling probability for unit i.
}
\examples{
#Generate data
set.seed(273)
data.dt <- generateStratifiedDomainEstimateData(n_samples = 100, generate_random_weights = FALSE)
head(data.dt)

#Get one of the variables and define sampling weights proportional to its size
data.dt <- data.dt[, c("income"), with = FALSE]

#Define sample size
n <- round(0.5 * nrow(data.dt))

#Variables with pi >= 1 have its selection probability set to 1
data.dt[, pi := n*income/sum(income)]
data.dt[, pi := ifelse(pi >= 1, 1, pi)]

#Define sampling weights
data.dt[, weights := 1/pi]
head(data.dt)

#Get sample proportional to sampling probabilities
data.keys <- 1:nrow(data.dt)
train.keys <- sample(x = data.keys, size = n, replace = FALSE, prob = data.dt[, pi])
test.keys <- setdiff(data.keys, train.keys)

target <- "income"
weights_column <- "weights"

#Define and train modeler
modeler <- HTRegressorModeler(target = target, weightsColumn = weights_column)

modeler <- train(object = modeler, data = data.dt, subSet = train.keys)

#Evaluate modeler
cf <- r2CostFunction(weightsColumn = "weights")

evaluate(object = modeler, data = data.dt, cost_function = cf, subSet = train.keys)
evaluate(object = modeler, data = data.dt, cost_function = cf, subSet = test.keys)


}
\seealso{
\code{\link{regressorModeler}}

\code{\link{strataHTRegressorModeler}}
}
