% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierModeler.R
\name{LGBClassifierModeler}
\alias{LGBClassifierModeler}
\alias{LGBClassifierModeler.list}
\title{Class constructor for LGBClassifierModeler}
\usage{
LGBClassifierModeler(params, target, regressors, positive_class = NULL)

\method{LGBClassifierModeler}{list}(params, target, regressors, positive_class = NULL)
}
\arguments{
\item{params}{List of parameters to be used for training a LightGBM model.
Tipically, the user may want to train a binary classifier or a multiclass
classifier for a binary variable, depending on their requirements
(as some parameters work only with one or the other). In that case, the
objective parameter should be either "binary" or "multiclass", respectively.}

\item{target}{String indicating the name of the target variable in the data.}

\item{regressors}{List or vector of strings indicating the regressor column names in the data.}

\item{positive_class}{String indicating the name of the positive class, only
used when params$objective is binary. Note that if this parameter is not
provided, it will be assumed that the target variable levels are the negative
class and the positive class, in that order.}
}
\value{
A LGBClassifierModeler object.
}
\description{
LightGBM implementation in rmlutils for classification tasks.
}
\examples{
#Generate data
data.dt <- generateBinaryClassificationData(n_samples = 1000,
                                            generate_random_weights = FALSE)

#Define train/test partitions
set.seed(273)
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

#Multiclass model
target <- "classification_target"
regressors <- setdiff(colnames(data.dt), target)
params <- list(
  "learning_rate" = 0.07,
  "nrounds" = 10,
  "objective" = "multiclass",
  "num_class" = length(unique(data.dt[, classification_target]))
)

#Define and train modeler
modeler <- LGBClassifierModeler(params = params,
                                target = target,
                                regressors = regressors)

modeler <- train(object = modeler,
                 data = data.dt,
                 subSet = train.keys)

#Get predictions and classes
preds <- getPredictions(object = modeler,
                        data = data.dt,
                        subSet = train.keys)

clss <- getClasses(object = modeler,
                   data = data.dt,
                   subSet = train.keys,
                   class_mapping = maxScoreClass())

#cf <- confusionMatrixCostFunction()
cf <- accuracyCostFunction()

#Obtain cost with getCost (this way we see the confusion matrix)
getCost(object = cf,
        data = data.dt,
        modeler = modeler,
        fSample = 1:nrow(data.dt),
        cSample = train.keys)

#Obtain cost with evaluate
evaluate(object = modeler,
         data = data.dt,
         cost_function = cf,
         subSet = test.keys)

#Binary model
target <- "classification_target"
regressors <- setdiff(colnames(data.dt), target)
params <- list(
  "learning_rate" = 0.07,
  "nrounds" = 10,
  "objective" = "binary"
)

#Specify positive class, otherwise, if target levels are not
#(negative, positive), it may not work properly
modeler2 <- LGBClassifierModeler(params = params,
                                 target = target,
                                 regressors = regressors,
                                 positive_class = "positive")

#Train and predict
modeler2 <- train(object = modeler2,
                  data = data.dt,
                  subSet = train.keys)

preds <- getPredictions(object = modeler2,
                        data = data.dt,
                        subSet = train.keys)

clss <- getClasses(object = modeler2,
                   data = data.dt,
                   subSet = train.keys,
                   class_mapping = maxScoreClass())

cf <- accuracyCostFunction()
#cf <- confusionMatrixCostFunction()

getCost(object = cf,
        data = data.dt,
        modeler = modeler2,
        fSample = 1:nrow(data.dt),
        cSample = train.keys)

evaluate(object = modeler2,
         data = data.dt,
         cost_function = cf,
         subSet = train.keys)


}
\seealso{
\code{\link{classifierModeler}}
}
