% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorModeler.R
\name{LGBRegressorModeler}
\alias{LGBRegressorModeler}
\alias{LGBRegressorModeler.list}
\title{Class constructor for LGBRegressorModeler}
\usage{
LGBRegressorModeler(params, target, regressors)

\method{LGBRegressorModeler}{list}(params, target, regressors)
}
\arguments{
\item{params}{List of parameters to be used for training a LightGBM model.}

\item{target}{String indicating the name of the target variable in the data.}

\item{regressors}{List or vector of strings indicating the regressor column names in the data.}
}
\value{
An LGBRegressorModeler object.
}
\description{
LightGBM implementation in rmlutils for regression tasks.
}
\examples{
#Generate data
data.dt <- generateRegressionData(n_samples = 1000, generate_random_weights = FALSE)

#Define train/test partitions
set.seed(273)
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

#Define modeler
params <- list(
  "learning_rate" = 0.07,
  "nrounds" = 10,
  "objective" = "regression"
)
target <- "regression_target"
regressors <- setdiff(colnames(data.dt), target)

modeler <- LGBRegressorModeler(params = params, target = target, regressors = regressors)

#Train, predict and evaluate
modeler <- train(modeler, data = data.dt, subSet = train.keys)

preds <- getPredictions(modeler, data = data.dt, subSet = train.keys)

cf <- r2CostFunction()

evaluate(object = modeler, data = data.dt, cost_function = cf, subSet = train.keys)

}
\seealso{
\code{\link{regressorModeler}}

\code{\link{LGBClassifierModeler}}
}
