% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMapping.R
\name{maxScoreClass}
\alias{maxScoreClass}
\title{Maximum score class mapping.}
\usage{
maxScoreClass(ties.method = "first")
}
\arguments{
\item{ties.method}{A string indicating the ties.method to use with max.col. Set to "first" by default. Additional options include "random" and "last".}
}
\value{
A classMapping object
}
\description{
Implementation of maximum score mapping. This serves as the default choice unless specified otherwise, by the user.
When given a matrix of scores for a set of samples, sending this object to the \code{\link{getClassMapping}} function will output an array indicating the position of the class with the highest score, for each sample.
}
\examples{

#Define class mapping (check getClassMapping and getClasses documentation to see a more complete example)
cls_mapping <- maxScoreClass(ties.method = "first")

}
\seealso{
\code{\link{getClassMapping}}

\code{\link{getClasses}}

\code{\link{binaryThresholdClass}}
}
