% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{MCSubRBDomainEstimate}
\alias{MCSubRBDomainEstimate}
\alias{MCSubRBDomainEstimate.regressorModeler}
\title{Monte Carlo Subsampling Rao-Blackwellized estimator by domains}
\usage{
MCSubRBDomainEstimate(
  auxModeler,
  data,
  mainSampler,
  subSampler,
  targetVar,
  domains,
  ssNumber = 50
)

\method{MCSubRBDomainEstimate}{regressorModeler}(
  auxModeler,
  data,
  mainSampler,
  subSampler,
  targetVar,
  domains,
  ssNumber = 50
)
}
\arguments{
\item{auxModeler}{A regressorModeler object to use for model-assisted
estimation on each subsample.}

\item{data}{A dataset containing the target variable of interest along with
other covariates used for model-assisted estimation.}

\item{mainSampler}{Sampler used to define the keys of the sub_sampler.
Its keys should match the sample_data indexes.}

\item{subSampler}{Sampler used to get ss_number samples from the target
variable. Note that its key should be defined from a main_sampler sample.}

\item{targetVar}{Name of the target variable in the dataset.}

\item{domains}{A list, vector, array or dataframe column indicating the
domains of interest. Its shape should match the rows of data.}

\item{ssNumber}{Amount of times to subsample.}
}
\value{
A list of lists containing the estimates and their variances for each domain, along their respective monte carlo errors.
}
\description{
Computes the Monte Carlo Subsampling Rao-Blackwellized estimator from
Sanguiao-Zhang (2020) (https://arxiv.org/abs/2003.11423) by domains. For a more
in-depth explanation, check \code{\link{MCSubRBEstimate}} documentation.
}
\examples{
#Generate data
set.seed(273)
data.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 1000, generate_random_weights = FALSE)
head(data.dt)

#Get relevant variables from data
data.keys <- 1:nrow(data.dt)
target <- "target"
regressors <- setdiff(names(data.dt), target)

#Define main sampler and sub-sampler
train_size = round(0.8 * nrow(data.dt))
main_sampler <- srsSampler(key = data.keys, ssize = train_size)
main_sample <- getSample(main_sampler)
sub_sampler <- srsSampler(key = main_sample, ssize = train_size*0.5)

#Define modeler
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)

modeler <- rangerRegressorModeler(params = params, target = target, regressors = regressors)

estim <- MCSubRBDomainEstimate(auxModeler = modeler,
                         data = data.dt,
                         mainSampler = main_sampler,
                         subSampler = sub_sampler,
                         targetVar = target,
                         domains = data.dt[, domain_class],
                         ssNumber = 100)

print(paste0("Real value of target: ", data.dt[, sum(target), by = c("domain_class")][, V1]))
print(paste0("Value of target estimate:", estim$total$value))
print(paste0("Monte Carlo error of estimate: ", estim$total$MC_error))
print(paste0("Standard error of estimate: ", sqrt(estim$variance$value)))
print(paste0("Standard error for the estimate of the Monte Carlo error: ", sqrt(estim$variance$MC_error)))

}
\seealso{
\code{\link{estimator}}

\code{\link{MCSubRBEstimate}}

\code{\link{biasVarianceEstimate}}
}
