% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{MCSubRBEstimate}
\alias{MCSubRBEstimate}
\alias{MCSubRBEstimate.regressorModeler}
\title{Monte Carlo Subsampling Rao-Blackwellized estimator.}
\usage{
MCSubRBEstimate(
  auxModeler,
  data,
  mainSampler,
  subSampler,
  targetVar,
  ssNumber = 50
)

\method{MCSubRBEstimate}{regressorModeler}(
  auxModeler,
  data,
  mainSampler,
  subSampler,
  targetVar,
  ssNumber = 50
)
}
\arguments{
\item{auxModeler}{A regressorModeler object to use for model-assisted
estimation on each subSample.}

\item{data}{A dataset containing the target variable of interest along with
other covariates used for model-assisted estimation.}

\item{mainSampler}{Sampler used to define the keys of the subSampler
Its keys should match the sample_data indexes.}

\item{subSampler}{Sampler used to get ssNumber samples from the target
variable. Note that its key should be defined from a mainSampler sample.}

\item{targetVar}{Name of the target variable in the dataset.}

\item{ssNumber}{Amount of times to sub-sample.}
}
\value{
A list of lists containing the estimate, its variance and the associated Monte Carlo errors.
}
\description{
Computes the Monte Carlo Subsampling Rao-Blackwellized estimator from
Sanguiao-Zhang (2020) (\url{https://arxiv.org/abs/2003.11423}).

Specifically, it makes use of equation (11) for model-assisted estimation.
Since exact Rao-Blackwellization is computationally expensive, a Monte Carlo
Subsampling method is used to obtain a reasonable estimation, which is
determined by the ssNumber parameter.

It proceeds as follows:

\enumerate{
 \item A total of ssNumber samples are generated by the subSampler
 \item For each subSample (s1):
 \itemize{
   \item A complementary sample to the subSample is defined (U/s1) (considering the mainSampler key as the full population).
   \item The auxiliary modeler is trained on the subSample data (s1).
   \item The auxiliary modeler predicts over the complementary data (U/s1).
   \item The modeler predictions (U/s1) and the true values in the subSample (s1) data are concatenated.
   \item A complementary sampler (cSampler) is defined, with its key set to be the complementary set of the current subSample for the total population (U/s1).
   \item A conditioned sample (cSample) is defined, as the set difference of the subSampler keys and the current subSample (s2).
   \item Errors are calculated comparing the modeler prediction and the real values (note that its shape matches the complementary sampler keys)(U/s1).
   \item The horvitz-thompson estimate for the sum of errors is calculated, assuming the conditioned_sampler as its sampling design and the conditioned sample (s2) as the subset seen by the HT modeler.
   \item Equation (11) is applied, adding up modeler predictions for the complementary subSample (U/s1), real values for the current subSample (s1), and estimate of total prediction errors for the conditioned sample (s2).
 }
 \item Finally, Monte Carlo errors are calculated for the total and variance estimates.
}
}
\examples{
#Generate data
set.seed(273)
data.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 1000, generate_random_weights = FALSE)
head(data.dt)

#Get relevant variables from data
data.keys <- 1:nrow(data.dt)
target <- "target"
regressors <- setdiff(names(data.dt), target)

#Define main sampler and sub-sampler
train_size = round(0.8 * nrow(data.dt))
main_sampler <- srsSampler(key = data.keys, ssize = train_size)
main_sample <- getSample(main_sampler)
sub_sampler <- srsSampler(key = main_sample, ssize = train_size*0.5)

#Define modeler
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)

modeler <- rangerRegressorModeler(params = params, target = target, regressors = regressors)

estim <- MCSubRBEstimate(auxModeler = modeler,
                         data = data.dt,
                         mainSampler = main_sampler,
                         subSampler = sub_sampler,
                         targetVar = target,
                         ssNumber = 100)

print(paste0("Real value of target: ", data.dt[, sum(target)]))
print(paste0("Value of target estimate:", estim$total$value))
print(paste0("Monte Carlo error of estimate: ", estim$total$MC_error))
print(paste0("Standard error of estimate: ", sqrt(estim$variance$value)))
print(paste0("Standard error for the estimate of the Monte Carlo error: ", sqrt(estim$variance$MC_error)))

}
\seealso{
\code{\link{estimator}}

\code{\link{MCSubRBDomainEstimate}}

\code{\link{biasVarianceEstimate}}
}
