% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costFunctions.R
\name{MCSubRBMSECostFunction}
\alias{MCSubRBMSECostFunction}
\alias{MCSubRBMSECostFunction.sampler}
\title{Mean squared error of the MCSubRBEstimator}
\usage{
MCSubRBMSECostFunction(sampler, sampleKeys, nonSampleData)

\method{MCSubRBMSECostFunction}{sampler}(sampler, sampleKeys, nonSampleData)
}
\arguments{
\item{sampler}{A sampler that has been used to generate a sample over the data.}

\item{sampleKeys}{Set of keys that was used to define the train data for a model.}

\item{nonSampleData}{Complementary data (test) that the model has not seen.}
}
\description{
A cost function that outputs the estimator, bias MSE and Monte Carlo error of
the estimator based on Zhang, L.-C., Sanguiao-Sande, L., & Lee, D.
(2025). Design-Based Predictive Inference. Journal of Official Statistics,
41(1), 404-432. \url{https://journals.sagepub.com/doi/10.1177/0282423X241277719}
It is the implementation of theorem 1.
}
\examples{
data.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 1000)

set.seed(123)

data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))

#Split train
sampler <- srsSampler(data.keys, ssize = train_size)

SKeys <- getSample(sampler)

train.dt <- data.dt[SKeys]

#Define modeler
target <- "target"
regressors <- setdiff(names(data.dt), target)
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)

regressor <- rangerRegressorModeler(params = params, target = target, regressors = regressors)
regressor <- train(regressor, data = train.dt)

#Define cost function and get cost
cf <- MCSubRBMSECostFunction(sampler, SKeys, nonSampleData = data.dt[-SKeys,])

crossValidate(regressor,
              data.dt[SKeys,],
              srsSampler(1:length(SKeys),150),#150 subsamples
              cf,
              times = 100)


}
\seealso{
\code{\link{costFunction}}

\code{\link{getCost}}
}
