% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorModeler.R
\name{memoryRegressorModeler}
\alias{memoryRegressorModeler}
\alias{memoryRegressorModeler.list}
\alias{memoryRegressorModeler.data.frame}
\title{Class constructor for memoryRegressorModeler}
\usage{
memoryRegressorModeler(data, modeler, maxData = 5L)

\method{memoryRegressorModeler}{list}(data, modeler, maxData = 5L)

\method{memoryRegressorModeler}{data.frame}(data, modeler, maxData = 5L)
}
\arguments{
\item{data}{list of data sets or a single data set}

\item{modeler}{a modeler object}

\item{maxData}{Optional value, maximum number of data sets in the memoryModeler}
}
\value{
A memoryRegressorModeler object.
}
\description{
This modeler saves together a list of data sets from different time instances (for example, years)
and a defined modeler. When training the modeler, all these data sets will be added to the data inputed
in the training.
If a single data set, instead of a list, is inputed, it will be considered as coming from a single time period.

It has additional methods to manage its internal data storage, namely:
\code{\link{addMemory}}, \code{\link{cleanMemory}} and \code{\link{updateMemory}}.
}
\examples{
#Generate data
set.seed(273)
data.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 1000, generate_random_weights = FALSE)
data2.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 100, generate_random_weights = FALSE)
data3.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 50, generate_random_weights = FALSE)

#Define train/test partitions
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

#Define base modeler
target <- "target"# Regression
regressors <- setdiff(names(data.dt), target)
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)
base_modeler <- rangerRegressorModeler(params = params,
                                       target = target,
                                       regressors = regressors)

#Instantiate memoryModeler
#memoryModeler parameters
data <- list(data2.dt)
maxData <- 3L
modeler <- memoryRegressorModeler(data = data,
                                  modeler = base_modeler,
                                  maxData = maxData)

#Train, getPredictions and evaluate
modeler <- train(modeler,
                 data = data.dt,
                 subSet = train.keys)

preds <- getPredictions(object = modeler,
                        data = data.dt,
                        subSet = test.keys)

#Define a cost function and evaluate
cf <- r2CostFunction()
#Evaluate over train set
evaluate(object = modeler,
         data = data.dt,
         cost_function = cf,
         subSet = train.keys)
#Evaluate over test set
evaluate(object = modeler,
         data = data.dt,
         cost_function = cf,
         subSet = test.keys)

}
\seealso{
\code{\link{addMemory}}

\code{\link{cleanMemory}}

\code{\link{updateMemory}}
}
