% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierModeler.R
\name{miceClassifierModeler}
\alias{miceClassifierModeler}
\alias{miceClassifierModeler.list}
\title{Class constructor for miceClassifierModeler}
\usage{
miceClassifierModeler(params, target, regressors)

\method{miceClassifierModeler}{list}(params, target, regressors = NULL)
}
\arguments{
\item{params}{List of parameters to be used for training a mice model.}

\item{target}{String indicating the name of the target variable in the data.}

\item{regressors}{List or vector of strings indicating the regressor column names in the data.}
}
\value{
A miceClassifierModeler object.
}
\description{
mice implementation in rmlutils for classification tasks.
Note that due to the way that mice works, its results are stochastic,
meaning that predicting over the same dataset multiple times will
usually provide different results. The user should use it with caution.
}
\examples{
#Generate data
#data.dt <- generateBinaryClassificationData(n_samples = 1000, generate_random_weights = FALSE)
data.dt <- generateMulticlassClassificationData(n_samples = 1000, generate_random_weights = FALSE)

#Define train/test partitions
set.seed(273)
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

target <- "classification_target"
regressors <- setdiff(colnames(data.dt), target)

#Define and train modeler
params <- list(m = 1,
               method = "rf",
               print = FALSE)

modeler <- miceClassifierModeler(params = params,
                                 target = target,
                                 regressors = regressors)

modeler <- train(object = modeler, data = data.dt, subSet = train.keys)

#Get predictions and classes
preds <- getPredictions(object = modeler,
                        data = data.dt,
                        subSet = train.keys)

clss <- getClasses(object = modeler,
                   data = data.dt,
                   subSet = train.keys,
                   class_mapping = maxScoreClass())

#cf <- confusionMatrixCostFunction()
cf <- accuracyCostFunction()

#Obtain cost with getCost (this way we see the confusion matrix)
getCost(object = cf,
        data = data.dt,
        modeler = modeler,
        fSample = 1:nrow(data.dt),
        cSample = train.keys)

#Obtain cost with evaluate
evaluate(object = modeler,
         data = data.dt,
         cost_function = cf,
         subSet = test.keys)


}
\seealso{
\code{\link{classifierModeler}}
}
