% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorModeler.R
\name{miceRegressorModeler}
\alias{miceRegressorModeler}
\alias{miceRegressorModeler.list}
\title{Class constructor for miceRegressorModeler}
\usage{
miceRegressorModeler(params, target, regressors)

\method{miceRegressorModeler}{list}(params, target, regressors = NULL)
}
\arguments{
\item{params}{List of parameters to be used for training a mice model.}

\item{target}{String indicating the name of the target variable in the data.}

\item{regressors}{List or vector of strings indicating the regressor column names in the data.}
}
\value{
A miceRegressorModeler object.
}
\description{
mice implementation in rmlutils for regression tasks.
Note that due to the way that mice works, its results are stochastic,
meaning that predicting over the same dataset multiple times will
usually provide different results. The user should use it with caution.
}
\examples{
#Generate data
data.dt <- generateRegressionData(n_samples = 1000, generate_random_weights = FALSE)

#Define train/test partitions
set.seed(273)
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

#Define modeler
params <- list(m = 1,
               method = "rf",
               print = FALSE)
target <- "regression_target"
regressors <- setdiff(colnames(data.dt), target)

modeler <- miceRegressorModeler(params = params, target = target, regressors = regressors)

#Train, predict and evaluate
modeler <- train(object = modeler, data = data.dt, subSet = train.keys)

preds <- getPredictions(modeler, data = data.dt, subSet = train.keys)

cf <- r2CostFunction()

evaluate(object = modeler, data = data.dt, cost_function = cf, subSet = train.keys)


}
\seealso{
\code{\link{regressorModeler}}

\code{\link{miceClassifierModeler}}
}
