% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{minMaxScalerPreproFunction}
\alias{minMaxScalerPreproFunction}
\title{A numeric min-max scaling preprocessing function}
\usage{
minMaxScalerPreproFunction(object = NULL)
}
\arguments{
\item{object}{\code{\link{preproObject}} to be used.
It should contain a preprocessing function and a set of parameters.

For this preprocessing function in particular, its parameters should be:

\itemize{
  \item modeler_type: String indicating the type of modeler that will
  be trained with preprocessed data. This is specifically to calculate
  the correct format for new_regressors. Its possible values are:
  \itemize{
     \item "simple" or NULL: A string vector will be assigned to new_regressors.
     \item "classifRegModeler": A named list with new_regressors for each
     internal modeler will be assigned.
     \item "prePredModeler": A named list with new_regressors for each
     internal modeler will be assigned.
     \item "benchRegressorModeler": An unnamed list with new_regressors for each
     internal modeler will be assigned.
  }
  \item pretrained_min_max (optional): A named list of lists, containing
  the minimum and maximum value for each numeric regressor. If left empty,
  it will be filled the first time that this function is called.
}

Note that this preprocessing function assumes that all internal modelers (for
complex modelers only) will be assigned the same regressors, except
prePredModeler, where its mainModeler will also have the predModeler target
in its new_regressors.}
}
\value{
\code{\link{preproObject}} with preprocessed data.
}
\description{
Given a \code{\link{preproObject}}, applies a Min-Max transformation to its
numeric data, so that it always ranges from 0 to 1.
Note that it saves the found minimum and maximum in the preproObject,
so that future executions will take those statistics into account,
instead of calculating them each time.
In general, the user should use the \code{\link{getPreproData}} function
to interact with it. Although it is recommended to use either a
\code{\link{preproRegressorModeler}} or \code{\link{preproClassifierModeler}}
instead of directly calling the previous method.

Prepro functions should only take the preproObject as an argument.
}
\examples{
#Generate data
data.dt <- generateRegressionData(n_samples = 1000,
                                  generate_random_weights = TRUE)
head(data.dt)

#Define relevant variables
target <- "regression_target"
weightsColumn <- "weights"
regressors <- setdiff(colnames(data.dt), c(target, weightsColumn))

modeler <- list("target" = target,
                "regressors" = regressors,
                "weightsColumn" = weightsColumn)

#Define preproObject
#Preprocessing parameters
prepro_params <- list(modeler_type = NULL, pretrained_min_max = NULL)

#Define preproObject
prepro_object <- preproObject(preproFunction = minMaxScalerPreproFunction,
                              params = prepro_params)

#Assign relevant metadata to preproObject
prepro_object <- assignExternalAttributes(modeler = modeler,
                                          object = prepro_object)

#Call getPreproData to obtain preproObject with preprocessed data
prepro_object <- getPreproData(object = prepro_object, data = data.dt)

#data is saved within prepro_object
head(prepro_object$data)

}
\seealso{
\code{\link{preproObject}}

\code{\link{assignExternalAttributes}}

\code{\link{getPreproData}}

\code{\link{preproClassifierModeler}}

\code{\link{preproRegressorModeler}}
}
