% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClean.R
\name{modelClean}
\alias{modelClean}
\alias{modelClean.default}
\alias{modelClean.modeler}
\alias{modelClean.directRegressorModeler}
\alias{modelClean.H2OModeler}
\alias{modelClean.preproModeler}
\alias{modelClean.classifRegModeler}
\alias{modelClean.prePredModeler}
\alias{modelClean.memoryModeler}
\alias{modelClean.benchModeler}
\title{Cleans a modeler object.}
\usage{
modelClean(object)

\method{modelClean}{default}(object)

\method{modelClean}{modeler}(object)

\method{modelClean}{directRegressorModeler}(object)

\method{modelClean}{H2OModeler}(object)

\method{modelClean}{preproModeler}(object)

\method{modelClean}{classifRegModeler}(object)

\method{modelClean}{prePredModeler}(object)

\method{modelClean}{memoryModeler}(object)

\method{modelClean}{benchModeler}(object)
}
\arguments{
\item{object}{Modeler to clean.}
}
\value{
The baseline model.
}
\description{
Removes any fit model from the modeler.
}
\examples{
#Generate data
data.dt <- generateRegressionData(n_samples = 1000,
                                  generate_random_weights = FALSE)

#Define train/test partitions
set.seed(273)
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

#Define modeler
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)
target <- "regression_target"
regressors <- setdiff(colnames(data.dt), target)

modeler <- rangerRegressorModeler(params = params, target = target, regressors = regressors)

#Train, predict and evaluate
modeler <- train(object = modeler, data = data.dt, subSet = train.keys)

print(modeler$isTrained)#Model is now trained

modeler <- modelClean(modeler)

print(modeler$isTrained)#Model is now clean

}
\seealso{
\code{\link{regressorModeler}}

\code{\link{classifierModeler}}

\code{\link{modelSave}}

\code{\link{modelLoad}}
}
