% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLoad.R
\name{modelLoad}
\alias{modelLoad}
\alias{modelLoad.character}
\alias{modelLoad.directModeler}
\alias{modelLoad.HTModeler}
\alias{modelLoad.strataHTModeler}
\alias{modelLoad.strataMeanModeler}
\alias{modelLoad.LGBModeler}
\alias{modelLoad.rangerModeler}
\alias{modelLoad.catboostModeler}
\alias{modelLoad.miceModeler}
\alias{modelLoad.H2OModeler}
\alias{modelLoad.classifRegModeler}
\alias{modelLoad.memoryModeler}
\alias{modelLoad.prePredModeler}
\alias{modelLoad.benchModeler}
\alias{modelLoad.preproModeler}
\title{Loads a modeler object (new version).}
\usage{
modelLoad(path, fileName, object = NULL)

\method{modelLoad}{character}(path, fileName, object = NULL)

\method{modelLoad}{directModeler}(object, path, fileName)

\method{modelLoad}{HTModeler}(object, path, fileName)

\method{modelLoad}{strataHTModeler}(object, path, fileName)

\method{modelLoad}{strataMeanModeler}(object, path, fileName)

\method{modelLoad}{LGBModeler}(object, path, fileName)

\method{modelLoad}{rangerModeler}(object, path, fileName)

\method{modelLoad}{catboostModeler}(object, path, fileName)

\method{modelLoad}{miceModeler}(object, path, fileName)

\method{modelLoad}{H2OModeler}(object, path, fileName)

\method{modelLoad}{classifRegModeler}(object, path, fileName)

\method{modelLoad}{memoryModeler}(object, path, fileName)

\method{modelLoad}{prePredModeler}(object, path, fileName)

\method{modelLoad}{benchModeler}(object, path, fileName)

\method{modelLoad}{preproModeler}(object, path, fileName)
}
\arguments{
\item{path}{path to the stored modeler object}

\item{fileName}{Name of the file where the modeler is stored.
If the fileName does not end with .mlu, this method will attempt to load
the modeler using an older version of modelLoad. It is recommended,
however, to load and save again older modelers, as such method will be
removed in the future. When saving older modelers, they will automatically
be converted to their new version.}
}
\description{
Loads a modeler object from disk.
}
\examples{
path <- "../../models/stored_models"
filename <- "my_model.mlu"

#mdl <- modelLoad(path, filename)

}
\seealso{
\code{\link{regressorModeler}}

\code{\link{classifierModeler}}

\code{\link{modelSave}}

\code{\link{modelClean}}
}
