% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLoadOld.R
\name{modelLoadOld}
\alias{modelLoadOld}
\alias{modelLoadOld.character}
\alias{modelLoadOld.directModeler}
\alias{modelLoadOld.HTModeler}
\alias{modelLoadOld.strataHTModeler}
\alias{modelLoadOld.strataMeanModeler}
\alias{modelLoadOld.benchModeler}
\alias{modelLoadOld.LGBModeler}
\alias{modelLoadOld.H2OModeler}
\alias{modelLoadOld.rangerModeler}
\alias{modelLoadOld.miceModeler}
\alias{modelLoadOld.catboostModeler}
\alias{modelLoadOld.preproModeler}
\alias{modelLoadOld.classifRegModeler}
\alias{modelLoadOld.prePredModeler}
\alias{modelLoadOld.memoryModeler}
\title{Loads a modeler object (old version, DO NOT USE).}
\usage{
modelLoadOld(path, fileName, object = NULL)

\method{modelLoadOld}{character}(path, fileName, object = NULL)

\method{modelLoadOld}{directModeler}(object, path, fileName)

\method{modelLoadOld}{HTModeler}(object, path, fileName)

\method{modelLoadOld}{strataHTModeler}(object, path, fileName)

\method{modelLoadOld}{strataMeanModeler}(object, path, fileName)

\method{modelLoadOld}{benchModeler}(object, path, fileName)

\method{modelLoadOld}{LGBModeler}(object, path, fileName)

\method{modelLoadOld}{H2OModeler}(object, path, fileName)

\method{modelLoadOld}{rangerModeler}(object, path, fileName)

\method{modelLoadOld}{miceModeler}(object, path, fileName)

\method{modelLoadOld}{catboostModeler}(object, path, fileName)

\method{modelLoadOld}{preproModeler}(object, path, fileName)

\method{modelLoadOld}{classifRegModeler}(object, path, fileName)

\method{modelLoadOld}{prePredModeler}(object, path, fileName)

\method{modelLoadOld}{memoryModeler}(object, path, fileName)
}
\arguments{
\item{path}{path to the stored modeler object}

\item{fileName}{Name of the file where the modeler is stored.}
}
\description{
Loads a modeler object from disk using an outdated method.
The user should avoid calling this function directly and use
modelLoad instead, as it will automatically handle older modelers without
having to call this one manually.
}
\examples{
path <- "../../models/stored_models"
filename <- "my_model"
#mdl <- modelLoadOld(path, filename)

}
