% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSave.R
\name{modelSave}
\alias{modelSave}
\alias{modelSave.directModeler}
\alias{modelSave.HTRegressorModeler}
\alias{modelSave.strataHTModeler}
\alias{modelSave.strataMeanModeler}
\alias{modelSave.benchModeler}
\alias{modelSave.H2OModeler}
\alias{modelSave.preproModeler}
\alias{modelSave.classifRegRegressorModeler}
\alias{modelSave.LGBModeler}
\alias{modelSave.rangerModeler}
\alias{modelSave.catboostModeler}
\alias{modelSave.miceModeler}
\alias{modelSave.prePredModeler}
\alias{modelSave.memoryModeler}
\title{Saves a modeler object.}
\usage{
modelSave(object, path, fileName)

\method{modelSave}{directModeler}(object, path, fileName)

\method{modelSave}{HTRegressorModeler}(object, path, fileName)

\method{modelSave}{strataHTModeler}(object, path, fileName)

\method{modelSave}{strataMeanModeler}(object, path, fileName)

\method{modelSave}{benchModeler}(object, path, fileName)

\method{modelSave}{H2OModeler}(object, path, fileName)

\method{modelSave}{preproModeler}(object, path, fileName)

\method{modelSave}{classifRegRegressorModeler}(object, path, fileName)

\method{modelSave}{LGBModeler}(object, path, fileName)

\method{modelSave}{rangerModeler}(object, path, fileName)

\method{modelSave}{catboostModeler}(object, path, fileName)

\method{modelSave}{miceModeler}(object, path, fileName)

\method{modelSave}{prePredModeler}(object, path, fileName)

\method{modelSave}{memoryModeler}(object, path, fileName)
}
\arguments{
\item{object}{\code{\link{modeler}} object to store}

\item{path}{Path to store the model}

\item{fileName}{Name under which save.}
}
\description{
Saves a modeler object to disk.
}
\examples{

#Modeler to save
modeler <- directRegressorModeler(protoTarget = "proto_target",
                                  scaleFactor = 1,
                                  target = "regression_target")

path <-  "../../stored_models"
filename <- "myModel"
#modelSave(modeler, path, filename)

}
\seealso{
\code{\link{regressorModeler}}

\code{\link{classifierModeler}}

\code{\link{modelLoad}}

\code{\link{modelClean}}
}
