% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorModeler.R
\name{prePredRegressorModeler}
\alias{prePredRegressorModeler}
\alias{prePredRegressorModeler.modeler}
\title{Class constructor for prePredRegressorModeler}
\usage{
prePredRegressorModeler(predModeler, mainModeler)

\method{prePredRegressorModeler}{modeler}(predModeler, mainModeler)
}
\arguments{
\item{predModeler}{A multivariate or univariate modeler to perform preliminary predictions for a set of regressor variables.}

\item{mainModeler}{The main modeler. The aforementioned predictions will be used as regressors, in addition to the ones already specified inside the main modeler.}
}
\value{
A prePredRegressorModeler object.
}
\description{
A two-step complex modeler. It works by first getting predictions for a set of variables, each having its own modeler object specified.
Thereupon, such predictions are implemented as regressors for a second modeler.
}
\examples{
#Generate data
data.dt <- generateRegressionData(n_samples = 1000, generate_random_weights = FALSE)
#Create an additional column that depends on regression_target
#This will be the actual target of this modeler
data.dt[, main_regression_target := regression_target + float_1 - float_2]

#Define train/test partitions
set.seed(273)
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

#Define previous modeler (this will predict regression_target)
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)
target <- "regression_target"
regressors <- setdiff(colnames(data.dt), c(target, "main_regression_target"))

predModeler <- rangerRegressorModeler(params = params,
                                      target = target,
                                      regressors = regressors)

#Define main modeler
#(this will predict main_regression_target,
#with regression_target as an additional regressor)
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)
target <- "main_regression_target"
regressors <- setdiff(colnames(data.dt), c(target))

mainModeler <- rangerRegressorModeler(params = params,
                                      target = target,
                                      regressors = regressors)

prePredModeler <- prePredRegressorModeler(predModeler = predModeler, mainModeler = mainModeler)

#Train, predict and evaluate
prePredModeler <- train(prePredModeler, data = data.dt, subSet = train.keys)

preds <- getPredictions(prePredModeler, data = data.dt, subSet = train.keys)

#Evaluate global output (mainModeler)
cf <- r2CostFunction()
evaluate(object = prePredModeler,
         data = data.dt,
         cost_function = cf,
         subSet = train.keys)

#Evaluate both internal modelers (predModeler and mainModeler)
cf2 <- list("predModeler" = r2CostFunction(),
            "mainModeler" = r2CostFunction())
evaluate(object = prePredModeler,
         data = data.dt,
         cost_function = cf2,
         subSet = train.keys)


}
\seealso{
\code{\link{regressorModeler}}
}
