% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierModeler.R
\name{preproClassifierModeler}
\alias{preproClassifierModeler}
\alias{preproClassifierModeler.modeler}
\title{Class constructor for preproClassifierModeler}
\usage{
preproClassifierModeler(modeler, preproObject)

\method{preproClassifierModeler}{modeler}(modeler, preproObject)
}
\arguments{
\item{modeler}{A classifierModeler object.}

\item{preproObject}{A \link{preproObject} structure, containing a preprocesing function to apply before training (or predicting) with the modeler.}
}
\value{
A preproClassifierModeler object.
}
\description{
preproModeler implementation for classification tasks.

A preproModeler applies a preprocessing function to the data before sending it to another modeler.
This pre-processing step will be applied right before training or predicting with the model.
}
\examples{
#data.dt <- generateBinaryClassificationData(n_samples = 1000, generate_random_weights = FALSE)
data.dt <- generateMulticlassClassificationData(n_samples = 1000, generate_random_weights = FALSE)

#Define train/test partitions
set.seed(273)
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

target <- "classification_target"
regressors <- setdiff(colnames(data.dt), target)

#Define and train modeler
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)

#Internal modeler
modeler <- rangerClassifierModeler(params = params,
                                   target = target,
                                   regressors = regressors)

#Preprocessing parameters
prepro_params <- list(modeler_type = NULL, pretrained_min_max = NULL)

#Define preproObject
prepro_object <- preproObject(preproFunction = minMaxScalerPreproFunction,
                              params = prepro_params)


p_modeler <- preproClassifierModeler(modeler = modeler,
                                     preproObject = prepro_object)

p_modeler <- train(object = p_modeler, data = data.dt, subSet = train.keys)

#Get predictions and classes
preds <- getPredictions(object = p_modeler,
                        data = data.dt,
                        subSet = train.keys)

clss <- getClasses(object = p_modeler,
                   data = data.dt,
                   subSet = train.keys,
                   class_mapping = maxScoreClass())

#cf <- confusionMatrixCostFunction()
cf <- accuracyCostFunction()

#Obtain cost with getCost (this way we see the confusion matrix)
getCost(object = cf,
        data = data.dt,
        modeler = p_modeler,
        fSample = 1:nrow(data.dt),
        cSample = train.keys)

#Obtain cost with evaluate
evaluate(object = p_modeler,
         data = data.dt,
         cost_function = cf,
         subSet = test.keys)


}
\seealso{
\code{\link{classifierModeler}}

\code{\link{preproRegressorModeler}}
}
