% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preproObject}
\alias{preproObject}
\alias{preproObject.function}
\title{Class constructor for preproObject class.}
\usage{
preproObject(preproFunction, params)

\method{preproObject}{`function`}(preproFunction, params)
}
\arguments{
\item{preproFunction}{A custom preprocessing function, such as \code{\link{minMaxScalerPreproFunction}} or any other function defined by the user (external to RMLutils).
These preproFunctions should only take a preproObject as their unique argument.}

\item{params}{A list of parameters to be used when calling preproFunction. The user is free to use them however they see fit.}
}
\value{
A preproObject.
}
\description{
Creates a preproObject as specified.

A preproObject is used to apply preprocessing steps before calling the train and getPredictions methods.
Once built, it can be used by calling the \link{getPreproData} method.
The user should generally use it mostly with preproModeler objects,
although it is possible to use it to preprocess data separately.
}
\examples{
#Generate data
data.dt <- generateRegressionData(n_samples = 1000,
                                  generate_random_weights = TRUE)
head(data.dt)

#Define relevant variables
target <- "regression_target"
weightsColumn <- "weights"
regressors <- setdiff(colnames(data.dt), c(target, weightsColumn))

modeler <- list("target" = target,
                "regressors" = regressors,
                "weightsColumn" = weightsColumn)

#Define preproObject
#Preprocessing parameters
prepro_params <- list(modeler_type = NULL, pretrained_min_max = NULL)

#Define preproObject
prepro_object <- preproObject(preproFunction = minMaxScalerPreproFunction,
                              params = prepro_params)

#Assign relevant metadata to preproObject
prepro_object <- assignExternalAttributes(modeler = modeler,
                                          object = prepro_object)

#Call getPreproData to obtain preproObject with preprocessed data
prepro_object <- getPreproData(object = prepro_object, data = data.dt)

#data is saved within prepro_object
head(prepro_object$data)

}
\seealso{
\code{\link{preproRegressorModeler}}

\code{\link{preproClassifierModeler}}

\code{\link{assignExternalAttributes}}

\code{\link{getPreproData}}
}
