% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorModeler.R
\name{preproRegressorModeler}
\alias{preproRegressorModeler}
\alias{preproRegressorModeler.modeler}
\title{Class constructor for preproRegressorModeler}
\usage{
preproRegressorModeler(modeler, preproObject)

\method{preproRegressorModeler}{modeler}(modeler, preproObject)
}
\arguments{
\item{modeler}{A regressorModeler object.}

\item{preproObject}{A \link{preproObject} structure, containing a preprocesing function to apply before training (or predicting) with the modeler.}
}
\value{
A preproRegressorModeler object.
}
\description{
preproModeler implementation for regression tasks.

A preproModeler applies a preprocessing function to the data before sending it to another modeler.
This pre-processing step will be applied right before training or predicting with the model.
}
\examples{
#Generate data
set.seed(273)
data.dt <- generateRegressionData(n_samples = 1000, generate_random_weights = FALSE)

#Define train/test partitions
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

target <- "regression_target"
regressors <- setdiff(colnames(data.dt), target)

#Preprocessing parameters
prepro_params <- list(pretrained_min_max = NULL)

#Define preproObject
prepro_object <- preproObject(preproFunction = minMaxScalerPreproFunction,
                              params = prepro_params)

#Define a regressor modeler
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)

modeler <- rangerRegressorModeler(params = params,
                                  target = target,
                                  regressors = regressors)

p_modeler <- preproRegressorModeler(modeler = modeler,
                                    preproObject = prepro_object)

p_modeler <- train(object = p_modeler, data = data.dt, subSet = train.keys)

cf <- r2CostFunction()

evaluate(object = p_modeler,
         data = data.dt,
         cost_function = cf,
         subSet = test.keys)

}
\seealso{
\code{\link{regressorModeler}}

\code{\link{preproClassifierModeler}}

\code{\link{preproObject}}

\code{\link{minMaxScalerPreproFunction}}
}
