% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{preTrainedDomainEstimate}
\alias{preTrainedDomainEstimate}
\alias{preTrainedDomainEstimate.regressorModeler}
\title{Pre-trained estimator by domains}
\usage{
preTrainedDomainEstimate(object, data, domain_column)

\method{preTrainedDomainEstimate}{regressorModeler}(object, data, domain_column)
}
\arguments{
\item{object}{Modeler object to make the estimation with.}

\item{data}{Dataset to make the estimation.}

\item{domain_column}{String indicating the domain column in the data.}
}
\value{
A vector with the estimates.
}
\description{
Generalization of preTrained estimator to estimate by domains.
}
\examples{
set.seed(273)
data.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 1000, generate_random_weights = FALSE)

#Get relevant variables from data
data.keys <- 1:nrow(data.dt)
target <- "target"# Regression
regressors <- setdiff(names(data.dt), target)

#Define sampler
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

#Define modeler
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)
modeler <- rangerRegressorModeler(params = params, target = target, regressors = regressors)

modeler <- train(modeler, data = data.dt, subSet = train.keys)

estim <- preTrainedDomainEstimate(modeler, data = data.dt, domain_column = "domain_class")
print(estim)
print(data.dt[, .(sum(target)), by = c("domain_class")])

}
\seealso{
\code{\link{estimator}}

\code{\link{preTrainedEstimate}}

\code{\link{getPredictions}}
}
