% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{preTrainedEstimate}
\alias{preTrainedEstimate}
\alias{preTrainedEstimate.regressorModeler}
\title{Pre-trained estimator}
\usage{
preTrainedEstimate(object, data)

\method{preTrainedEstimate}{regressorModeler}(object, data)
}
\arguments{
\item{object}{Modeler object to make the estimation with.}

\item{data}{Dataset to make the estimation.}
}
\value{
A vector with the estimates.
}
\description{
Computes the prediction estimation from a pre-trained model,
assuming that the model is capable of outputting a reasonably unbiased estimate
of the target variable.
}
\examples{
set.seed(273)
data.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 1000, generate_random_weights = FALSE)

#Get relevant variables from data
data.keys <- 1:nrow(data.dt)
target <- "target"# Regression
regressors <- setdiff(names(data.dt), target)

#Define sampler
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

#Define modeler
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)
modeler <- rangerRegressorModeler(params = params, target = target, regressors = regressors)

modeler <- train(modeler, data = data.dt, subSet = train.keys)

estim <- preTrainedEstimate(modeler, data = data.dt)
print(estim)
print(data.dt[, sum(target)])

}
\seealso{
\code{\link{estimator}}

\code{\link{preTrainedDomainEstimate}}

\code{\link{getPredictions}}
}
