% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierModeler.R
\name{rangerClassifierModeler}
\alias{rangerClassifierModeler}
\alias{rangerClassifierModeler.list}
\title{Class constructor for rangerClassifierModeler}
\usage{
rangerClassifierModeler(params, target, regressors)

\method{rangerClassifierModeler}{list}(params, target, regressors)
}
\arguments{
\item{params}{List of parameters to be used for training a ranger model.}

\item{target}{String indicating the name of the target variable in the data.}

\item{regressors}{List or vector of strings indicating the regressor column names in the data.}
}
\value{
A rangerClassifierModeler object.
}
\description{
ranger implementation in rmlutils for classification tasks.

Wright, M. N. & Ziegler, A. (2017). ranger: A fast implementation of random
forests for high dimensional data in C++ and R. J Stat Softw
77:1-17. \url{https://doi.org/10.18637/jss.v077.i01}.
}
\examples{
#Generate data
#data.dt <- generateBinaryClassificationData(n_samples = 1000, generate_random_weights = FALSE)
data.dt <- generateMulticlassClassificationData(n_samples = 1000, generate_random_weights = FALSE)

#Define train/test partitions
set.seed(273)
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

target <- "classification_target"
regressors <- setdiff(colnames(data.dt), target)

#Define and train modeler
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)

modeler <- rangerClassifierModeler(params = params,
                                   target = target,
                                   regressors = regressors)

modeler <- train(object = modeler, data = data.dt, subSet = train.keys)

#Get predictions and classes
preds <- getPredictions(object = modeler,
                        data = data.dt,
                        subSet = train.keys)

clss <- getClasses(object = modeler,
                   data = data.dt,
                   subSet = train.keys,
                   class_mapping = maxScoreClass())

#cf <- confusionMatrixCostFunction()
cf <- accuracyCostFunction()

#Obtain cost with getCost (this way we see the confusion matrix)
getCost(object = cf,
        data = data.dt,
        modeler = modeler,
        fSample = 1:nrow(data.dt),
        cSample = train.keys)

#Obtain cost with evaluate
evaluate(object = modeler,
         data = data.dt,
         cost_function = cf,
         subSet = test.keys)


}
\seealso{
\code{\link{classifierModeler}}
}
