% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorModeler.R
\name{rangerRegressorModeler}
\alias{rangerRegressorModeler}
\alias{rangerRegressorModeler.list}
\title{Class constructor for rangerRegressorModeler.}
\usage{
rangerRegressorModeler(params, target, regressors)

\method{rangerRegressorModeler}{list}(params, target, regressors)
}
\arguments{
\item{params}{List of parameters to be used for training a ranger model.}

\item{target}{String indicating the name of the target variable in the data.}

\item{regressors}{List or vector of strings indicating the regressor column names in the data.}
}
\value{
A rangerRegressorModeler object.
}
\description{
ranger implementation in rmlutils for regression tasks.

Wright, M. N. & Ziegler, A. (2017). ranger: A fast implementation of random
forests for high dimensional data in C++ and R. J Stat Softw
77:1-17. \url{https://doi.org/10.18637/jss.v077.i01}.
}
\examples{
#Generate data
data.dt <- generateRegressionData(n_samples = 1000, generate_random_weights = FALSE)

#Define train/test partitions
set.seed(273)
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

#Define modeler
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)
target <- "regression_target"
regressors <- setdiff(colnames(data.dt), target)

modeler <- rangerRegressorModeler(params = params, target = target, regressors = regressors)

#Train, predict and evaluate
modeler <- train(object = modeler, data = data.dt, subSet = train.keys)

preds <- getPredictions(modeler, data = data.dt, subSet = train.keys)

cf <- r2CostFunction()

evaluate(object = modeler, data = data.dt, cost_function = cf, subSet = train.keys)

}
\seealso{
\code{\link{regressorModeler}}

\code{\link{rangerClassifierModeler}}
}
