% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{RDomainEstimate}
\alias{RDomainEstimate}
\alias{RDomainEstimate.default}
\title{Raulin's formula for variance estimation by domain.}
\usage{
RDomainEstimate(data, strata_colummn, domain_column, weights_column)

\method{RDomainEstimate}{default}(data, strata_colummn, domain_column, weights_column)
}
\arguments{
\item{data}{Dataset of information and variables of the sampled Data.}

\item{strata_colummn}{String indicating the column name that identifies strata.}

\item{weights_column}{String indicating the column name that contains sampling weights.}

\item{domain_colummn}{String indicating the column name that identifies domains (subpopulations of interest).}
}
\value{
A list of lists, with the Horvitz-Thompson estimation of the total and its
variance estimation for each domain in the data.
}
\description{
Computes the Horvitz-Thompson estimator using Raulin's
formula for variance when data has multiple subpopulations (domains) of
interest (assuming stratified random sampling without replacement):

\deqn{
 \hat{V}(\hat{Y}_d) = \sum_{h=1}^{H} \frac{n_h}{n_h-1} \sum_{i \in h} w_i(w_i-1) (y_{i,d} - \bar{y}_{h,d})^2
}

where:
\itemize{
  \item \eqn{H}: Total amount of strata.
  \item \eqn{n_h}: sample size in each strata \eqn{h}
  \item \eqn{y_{i,d} = y_i \cdot I(i \in d)}: value of the target variable adjusted by the domain belonging indicator.
  \item \eqn{\bar{y}_{h,d} = \frac{1}{n_h} \sum_{i \in h} y_{i,d}}: unweighted average of the domain within the stratum \eqn{h}
  \item \eqn{w_i}: sampling weight for unit \eqn{i}
}

All numeric (and non-parameter) columns in the data are assumed to be target variables to compute the estimate.
}
\examples{
data.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 1000,
                                                             generate_random_weights = FALSE)
head(data.dt)
data.dt[, target := NULL]
#Define stratified random sampler proportional to sample size
set.seed(273)
key = 1:nrow(data.dt)
sssmplr <- ssSampler(key = key,
                     strata = data.dt[, strata_class],
                     ssize = c("Rural" = 408, "Urban" = 392))
sample_2 <- getSample(sssmplr)
#Define weights column
#Since sampling probabilities are pi = n_h/N_h
#Weights are 1/pi
data.dt[, N_h := .N , by = c("strata_class")]
#Vector of n_h
strata_to_sample_size <- c("Rural" = 408, "Urban" = 392)
data.dt[, weights := N_h/strata_to_sample_size[strata_class]]
data.dt[, N_h := NULL]

strata_colummn <- "strata_class"
domain_column <- "domain_column"
weights_column <- "weights"

estim <- RDomainEstimate(data = data.dt[sample_2,],
                         strata_colummn = strata_colummn,
                         domain_column = domain_column,
                         weights_column = weights_column)
print(estim)
print(paste0("Real total value of income: ", data.dt[, sum(income)]))
print(paste0("Real total value of cost_of_living: ", data.dt[, sum(cost_of_living)]))

}
\seealso{
\code{\link{estimator}}

\code{\link{REstimate}}
}
