% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorModeler.R
\name{regressorModeler}
\alias{regressorModeler}
\title{Class constructors for regression modeler classes.}
\value{
A regressorModeler object.
}
\description{
Creates a modeler object as specified.

The modeler classes contain different types of machine learning models to
provide a common interface for all of them.
}
\section{Current implementations}{


\strong{Simple modelers:}
\itemize{
\item \code{\link{directRegressorModeler}}
\item \code{\link{HTRegressorModeler}}
\item \code{\link{strataHTRegressorModeler}}
\item \code{\link{strataMeanRegressorModeler}}
\item \code{\link{H2ORegressorModeler}}
\item \code{\link{LGBRegressorModeler}}
\item \code{\link{rangerRegressorModeler}}
\item \code{\link{catboostRegressorModeler}}
\item \code{\link{miceRegressorModeler}}
}

\strong{Complex modelers:}
\itemize{
\item \code{\link{classifRegRegressorModeler}}
\item \code{\link{benchRegressorModeler}}
\item \code{\link{preproRegressorModeler}}
\item \code{\link{prePredRegressorModeler}}
\item \code{\link{memoryRegressorModeler}}
}
}

