% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{REstimate}
\alias{REstimate}
\alias{REstimate.default}
\title{Raulin's formula for variance estimation.}
\usage{
REstimate(data, strata_colummn, weights_column)

\method{REstimate}{default}(data = NULL, strata_colummn = NULL, weights_column = NULL)
}
\arguments{
\item{data}{Dataset of information and variables of the sampled Data.}

\item{strata_colummn}{String indicating the column name that identifies strata.}

\item{weights_column}{String indicating the column name that contains sampling weights.}
}
\value{
A list with the Horvitz-Thompson estimation of the total and its
variance estimation.
}
\description{
Computes the Horvitz-Thompson estimator using Raulin's
formula for variance (assuming stratified random sampling without replacement).

It is an approximation of variance that only requires the first-order selection
probabilities of a sample for obtaining it.

\deqn{
 \hat{V}(\hat{Y}) = \sum_{h=1}^{H} \frac{n_h}{n_h-1} \sum_{i \in h} w_i(w_i-1) (y_i - \bar{y}_h)^2
}

where:
\itemize{
  \item \eqn{H}: Total amount of strata.
  \item \eqn{n_h}: sample size in each strata \eqn{h}
  \item \eqn{y_{i} = y_i}: value of the target variable.
  \item \eqn{\bar{y}_{h} = \frac{1}{n_h} \sum_{i \in h} y_{i}}: unweighted average of the target variable within the stratum \eqn{h}
  \item \eqn{w_i}: sampling weight for unit \eqn{i}
}

All numeric (and non-parameter) columns in the data are assumed to be target variables to compute the estimate.
}
\examples{
#Generate stratified data
data.dt <- generateStratifiedEstimateData(n_samples = 1000,
                                          generate_random_weights = FALSE)
#Define stratified random sampler proportional to sample size
set.seed(273)
key = 1:nrow(data.dt)
sssmplr <- ssSampler(key = key,
                     strata = data.dt[, strata_class],
                     ssize = c("Class_1" = 400, "Class_2" = 240, "Class_3" = 160))
sample_2 <- getSample(sssmplr)

#Define weights column
#Since sampling probabilities are pi = n_h/N_h
#Weights are 1/pi
data.dt[, N_h := .N , by = c("strata_class")]
#Vector of n_h
strata_to_sample_size <- c("Class_1" = 400, "Class_2" = 240, "Class_3" = 160)
data.dt[, weights := N_h/strata_to_sample_size[strata_class]]
data.dt[, N_h := NULL]
strata_colummn <- "strata_class"
weights_column <- "weights"
estim <- REstimate(data = data.dt[sample_2],
                   strata_colummn = strata_colummn,
                   weights_column = weights_column)
print(estim)

#Since correct valies are known (see help(generateStratifiedEstimateData) for more details)
c("Target_1" = 500 * 10 + 300 * 1 + 200 * 20,
  "Target_2" = 500 * (-20) + 300 * (-20) + 200 * (-20),
  "Target_3" = 500 * 0 + 300 * 30 + 200 * 10)

}
\seealso{
\code{\link{estimator}}

\code{\link{RDomainEstimate}}
}
