% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costFunctions.R
\name{ROCAUCCostFunction}
\alias{ROCAUCCostFunction}
\title{Receiver Operating Characteristic Area Under Curve (ROC AUC) cost function}
\usage{
ROCAUCCostFunction(weightsColumn = "", average = "micro")
}
\arguments{
\item{weightsColumn}{Name of the weights column in the data}

\item{average}{String indicating the averaging method. This will generally only affect the multiclass case. Possible values are:

\itemize{
 \item "micro": Global metric average (weighted by sample weights, if provided).
 \item "macro": Average class metric (weighted by sample weights, if provided).
 \item "weighted": Average class metric (weighted by sample weights, if provided, and number of times that each class appears in the dataset).
}}
}
\description{
A costFunction object that outputs the ROC-AUC score as a performance metric for a \code{\link{classifierModeler}}.
Like any other cost function, it must be used together with the \code{\link{getCost}} function to output its metric.
}
\examples{
#Generate data
data.dt <- generateMulticlassClassificationData(n_samples = 1000)
#Instantiate cost function
cf <- ROCAUCCostFunction(weightsColumn = "", average = "micro")

#Train simple modeler
target <- "classification_target"
regressors <- setdiff(names(data.dt), target)
params <- list(
  "learning_rate" = 0.07,
  "shrinkage_rate" = 0.12,
  "nrounds" = 10,#Avoid warning
  "objective" = "multiclass",#Works better this way with RMLutils
  "num_class" = length(unique(data.dt[, classification_target]))
)

classifier <- LGBClassifierModeler(params = params, target = target, regressors = regressors)
classifier <- train(classifier, data = data.dt)

#Evaluate with getCost
cost <- getCost(object = cf, data = data.dt, modeler = classifier, cSample = 1:nrow(data.dt))
print(cost$classification_target$cost)
head(cost$classification_target$class_metrics$tpr)
head(cost$classification_target$class_metrics$fpr)

}
\seealso{
\code{\link{costFunction}}

\code{\link{getCost}}
}
