% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_alt}
\alias{sample_alt}
\title{Alternate sample function}
\usage{
sample_alt(...)
}
\arguments{
\item{...}{Common arguments to send to \code{\link{sample}}.
\itemize{
  \item \code{x}: vector of elements to sample from.
  \item \code{size}: amount of elements to sample.
  \item \code{replace}: logical indicating whether to allow for replacement when sampling.
  \item \code{prob}: vector with sampling probabilities.
}}
}
\value{
A vector of length \code{size} with the sampled elements.
  If \code{x} is of length 1, returns \code{rep(x, size)}.
  Otherwise, it is identical to \code{sample}.
}
\description{
A custom variant to the base R function sample that allows samples from single-element vectors.
}
\examples{
set.seed(273)
# Single value vector to sample from
sample_alt(x = c(1), size = 3)

#Normal case
sample_alt(x = 1:10, size = 5)


}
\seealso{
\code{\link{sample}}
}
