% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler}
\alias{sampler}
\title{A collection of class constructors for several sampler classes.}
\value{
A sampler object.
}
\description{
Creates a sampler object as specified.

The sampler classes contain different types of sampling schemes to
provide a common interface for all of them.
}
\section{Current implementations}{


\itemize{
\item \code{\link{srsSampler}}
\item \code{\link{ssSampler}}
\item \code{\link{bootstrapSampler}}
\item \code{\link{fullSampler}}
\item \code{\link{cvSampler}}
\item \code{\link{cvGroupSampler}}
\item \code{\link{cvMixedSampler}}
\item \code{\link{cvTemporalSampler}}
}
}

