% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRegressors.R
\name{setRegressors}
\alias{setRegressors}
\alias{setRegressors.default}
\alias{setRegressors.modeler}
\alias{setRegressors.preproModeler}
\alias{setRegressors.classifRegModeler}
\alias{setRegressors.memoryRegressorModeler}
\alias{setRegressors.prePredModeler}
\alias{setRegressors.benchRegressorModeler}
\title{Sets regressors attribute in a modeler}
\usage{
setRegressors(object, newRegressors)

\method{setRegressors}{default}(object, newRegressors)

\method{setRegressors}{modeler}(object, newRegressors)

\method{setRegressors}{preproModeler}(object, newRegressors)

\method{setRegressors}{classifRegModeler}(object, newRegressors)

\method{setRegressors}{memoryRegressorModeler}(object, newRegressors)

\method{setRegressors}{prePredModeler}(object, newRegressors)

\method{setRegressors}{benchRegressorModeler}(object, newRegressors)
}
\arguments{
\item{object}{A modeler object.}

\item{newRegressors}{String vector or list of string vectors.}
}
\value{
A modeler object with its regressors attribute changed.
}
\description{
Sets regressors into a modeler.
}
\examples{
mdl <- H2ORegressorModeler("glm", list(lambda=0.0),"d" , c("a", "b", "c"))
mdl <- setRegressors(mdl, c("e", "f", "g"))
getRegressors(mdl)

}
