% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTarget.R
\name{setTarget}
\alias{setTarget}
\alias{setTarget.default}
\alias{setTarget.modeler}
\alias{setTarget.preproModeler}
\alias{setTarget.classifRegModeler}
\alias{setTarget.memoryRegressorModeler}
\alias{setTarget.prePredModeler}
\alias{setTarget.benchRegressorModeler}
\title{Sets target attribute in a modeler}
\usage{
setTarget(object, newTarget)

\method{setTarget}{default}(object, newTarget)

\method{setTarget}{modeler}(object, newTarget)

\method{setTarget}{preproModeler}(object, newTarget)

\method{setTarget}{classifRegModeler}(object, newTarget)

\method{setTarget}{memoryRegressorModeler}(object, newTarget)

\method{setTarget}{prePredModeler}(object, newTarget)

\method{setTarget}{benchRegressorModeler}(object, newTarget)
}
\arguments{
\item{object}{A modeler object.}

\item{newTarget}{String or list with the new target name(s).}
}
\value{
A modeler object with its target attribute changed.
}
\description{
Sets one or multiple target variables into a modeler.
}
\examples{
mdl <- H2ORegressorModeler("glm", list(lambda=0.0),"d" , c("a", "b", "c"))
mdl <- setTarget(mdl, "e")
getTarget(mdl)

}
