% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costFunctions.R
\name{squareBiasCostFunction}
\alias{squareBiasCostFunction}
\title{Squared bias cost function}
\usage{
squareBiasCostFunction(weightsColumn = "")
}
\arguments{
\item{weightsColumn}{Name of the weights column in the data. Set as "" by default}
}
\description{
A costFunction object that outputs the squared bias as a performance metric for a \code{\link{regressorModeler}}.
Like any other cost function, it must be used together with the \code{\link{getCost}} function to output its metric.
}
\examples{
#Generate data
data.dt <- generateRegressionData(n_samples = 1000, generate_random_weights = FALSE)

#Instantiate cost function
cf <- squareBiasCostFunction(weightsColumn = "")

target <- "regression_target"# Regression

regressors <- setdiff(names(data.dt), target)

#Define modeler
params <- list(
  "learning_rate" = 0.07,
  "nrounds" = 10,
  "objective" = "regression"
)

regressor <- LGBRegressorModeler(params = params, target = target, regressors = regressors)
regressor <- train(regressor, data = data.dt)

cost <- getCost(object = cf,
                data = data.dt,
                modeler = regressor,
                fSample = 1:nrow(data.dt),
                cSample = 1:nrow(data.dt))
print(cost)

}
\seealso{
\code{\link{costFunction}}

\code{\link{getCost}}
}
