% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{srsSampler}
\alias{srsSampler}
\title{Simple random sampler}
\usage{
srsSampler(key, ssize)
}
\arguments{
\item{key}{Vector of indices of the data entries (rows) to sample from.}

\item{ssize}{Amount of samples to be taken.}
}
\value{
A srsSampler object.
}
\description{
Implementation of simple random sampling without replacement.
}
\examples{
set.seed(273)
#Define data keys
sample.keys <- 1:15
#Define sampler
sampler <- srsSampler(key = sample.keys, ssize = 10)
#Obtain a sample
getSample(sampler)

}
\seealso{
\code{\link{sampler}}

\code{\link{ssSampler}}

\code{\link{bootstrapSampler}}

\code{\link{cvSampler}}

\code{\link{fullSampler}}
}
