% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{ssSampler}
\alias{ssSampler}
\title{Stratified simple random sampler}
\usage{
ssSampler(key, strata, ssize)
}
\arguments{
\item{key}{Vector of indices of the data entries (rows) to sample from.}

\item{strata}{List of strata corresponding to each of the entries to sample from.}

\item{ssize}{Named vector indicating the amount of samples to take from each strata.}
}
\value{
A ssSampler object.
}
\description{
Implementation of stratified random sampling without replacement. The user must manually specify how many units to sample from each strata.
}
\examples{
set.seed(273)
#Generate data and define keys
#This data has two strata: Urban and Rural
data.dt <- generateStratifiedDomainEstimateData(n_samples = 100)
key <- 1:nrow(data.dt)
head(data.dt)
smplr <- ssSampler(key = key,
                   strata = data.dt[, strata_class],
                   ssize = c("Urban" = 20, "Rural" = 15))
s1 <- getSample(smplr)
print(table(data.dt[s1, strata_class]))
print(data.dt[s1])


}
\seealso{
\code{\link{sampler}}

\code{\link{srsSampler}}

\code{\link{bootstrapSampler}}

\code{\link{cvSampler}}

\code{\link{fullSampler}}
}
