% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorModeler.R
\name{strataHTRegressorModeler}
\alias{strataHTRegressorModeler}
\alias{strataHTRegressorModeler.character}
\title{Class constructor for strataHTRegressorModeler}
\usage{
strataHTRegressorModeler(weightsColumn, target, strata)

\method{strataHTRegressorModeler}{character}(weightsColumn, target, strata)
}
\arguments{
\item{weightsColumn}{Name of the weights column.}

\item{target}{Name of the target variable.}

\item{strata}{Name of the strata information column.}
}
\value{
An strataHTRegressorModeler object.
}
\description{
An alternate version of \code{\link{HTRegressorModeler}} for
stratified data.
}
\examples{
#Generate data
set.seed(273)
data.dt <- generateStratifiedDomainEstimateData(n_samples = 100, generate_random_weights = FALSE)

#Get one of the variables and define sampling weights proportional to its size
data.dt <- data.dt[, c("income", "strata_class"), with = FALSE]

#Define sample size
n <- round(0.5 * nrow(data.dt))

#Variables with pi >= 1 have its selection probability set to 1
data.dt[, pi := n*income/sum(income)]
data.dt[, pi := ifelse(pi >= 1, 1, pi)]

#Define sampling weights
data.dt[, weights := 1/pi]
head(data.dt)

#Get sample proportional to sampling probabilities
data.keys <- 1:nrow(data.dt)
train.keys <- sample(x = data.keys, size = n, replace = FALSE, prob = data.dt[, pi])
test.keys <- setdiff(data.keys, train.keys)

target <- "income"
weights_column <- "weights"
strata_column <- "strata_class"

#Define and train modeler
modeler <- strataHTRegressorModeler(weightsColumn = weights_column,
                                    target = target,
                                    strata = strata_column)

modeler <- train(object = modeler, data = data.dt, subSet = train.keys)

#Evaluate modeler
cf <- r2CostFunction(weightsColumn = "weights")

evaluate(object = modeler, data = data.dt, cost_function = cf, subSet = train.keys)
evaluate(object = modeler, data = data.dt, cost_function = cf, subSet = test.keys)

}
\seealso{
\code{\link{regressorModeler}}

\code{\link{HTRegressorModeler}}
}
