% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorModeler.R
\name{strataMeanRegressorModeler}
\alias{strataMeanRegressorModeler}
\alias{strataMeanRegressorModeler.character}
\title{Class constructor for strataMeanRegressorModeler.}
\usage{
strataMeanRegressorModeler(target, regressors, weightsColumn)

\method{strataMeanRegressorModeler}{character}(target, regressors, weightsColumn = "")
}
\arguments{
\item{target}{Name of the target variable.}

\item{regressors}{Name or list of names of the variables to use as regressors.
Note that these variables must be categorical or of character type to
properly define strata.}

\item{weightsColumn}{Name of the weights column.}
}
\value{
A strataMeanRegressorModeler object.
}
\description{
A regressor modeler that predicts a target variable by
calculating its average over several regressors used as stratifying variables.
}
\examples{
set.seed(273)
data.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 100,
                                                             generate_random_weights = TRUE)
#Define sampler
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

#Define modeler parameters
target <- "target"# Regression
weightsColumn <- "weights"
regressors <- c("strata_class", "domain_class")#Only categorical variables

#Define and train modeler
modeler <- strataMeanRegressorModeler(target = target,
                                      regressors = regressors,
                                      weightsColumn = "weights")

modeler <- train(object = modeler, data = data.dt, subSet = train.keys)

#Evaluate modeler
cf <- MSECostFunction(weightsColumn = "weights")

evaluate(object = modeler, data = data.dt,
         cost_function = cf, subSet = train.keys)
evaluate(object = modeler, data = data.dt,
         cost_function = cf, subSet = test.keys)


}
\seealso{
\code{\link{regressorModeler}}
}
