% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\alias{train.directModeler}
\alias{train.HTModeler}
\alias{train.strataHTModeler}
\alias{train.strataMeanModeler}
\alias{train.H2OModeler}
\alias{train.benchModeler}
\alias{train.preproModeler}
\alias{train.LGBModeler}
\alias{train.rangerModeler}
\alias{train.catboostModeler}
\alias{train.miceModeler}
\alias{train.prePredModeler}
\alias{train.classifRegModeler}
\alias{train.memoryModeler}
\title{Trains a modeler object.}
\usage{
train(object, data, subSet = NULL)

\method{train}{directModeler}(object, data, subSet = NULL)

\method{train}{HTModeler}(object, data, subSet = NULL)

\method{train}{strataHTModeler}(object, data, subSet = NULL)

\method{train}{strataMeanModeler}(object, data, subSet = NULL)

\method{train}{H2OModeler}(object, data, subSet = NULL)

\method{train}{benchModeler}(object, data, subSet = NULL)

\method{train}{preproModeler}(object, data, subSet = NULL)

\method{train}{LGBModeler}(object, data, subSet = NULL)

\method{train}{rangerModeler}(object, data, subSet = NULL)

\method{train}{catboostModeler}(object, data, subSet = NULL)

\method{train}{miceModeler}(object, data, subSet = NULL)

\method{train}{prePredModeler}(object, data, subSet = NULL)

\method{train}{classifRegModeler}(object, data, subSet = NULL)

\method{train}{memoryModeler}(object, data, subSet = NULL)
}
\arguments{
\item{object}{\code{\link{modeler}} object to fit to the data.}

\item{data}{Dataframe containing the data to use for the training}

\item{subSet}{Optional value, select a subSet of the data for the training.}

\item{val_data}{Optional value for catbostModeler objects. Used as validation set during training.}
}
\value{
A trained modeler object, which contains some of the details and metrics
after training.
}
\description{
Fits the model in the modeler generated by \code{\link{modeler}} container to the given data.
}
\examples{
#Generate data
data.dt <- generateRegressionData(n_samples = 1000, generate_random_weights = FALSE)

#Define train/test partitions
set.seed(273)
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

#Define modeler
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)
target <- "regression_target"
regressors <- setdiff(colnames(data.dt), target)

modeler <- rangerRegressorModeler(params = params, target = target, regressors = regressors)

#Train modeler
modeler <- train(object = modeler, data = data.dt, subSet = train.keys)
}
