% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateMemory}
\alias{updateMemory}
\alias{updateMemory.memoryModeler}
\title{update new data into memoryModeler}
\usage{
updateMemory(object, data, index = 1)

\method{updateMemory}{memoryModeler}(object, data, index = 1)
}
\arguments{
\item{object}{\code{\link{modeler}} object to fit to the data.}

\item{data}{Dataframe containing the data to use for the training}

\item{index}{Optional value, select the position of the data to be updated.
By default, index = 1 (newest data)}
}
\value{
memoryModeler object with updated data.
}
\description{
update new data to memoryModeler into the specified position, substituting the original data
}
\examples{
#Generate data
set.seed(273)
data.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 1000, generate_random_weights = FALSE)
data2.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 100, generate_random_weights = FALSE)
data3.dt <- generateStratifiedDomainModelAssistedEstimateData(n_samples = 50, generate_random_weights = FALSE)

#Define train/test partitions
data.keys <- 1:nrow(data.dt)
train_size = round(0.8 * nrow(data.dt))
smplr <- srsSampler(key = data.keys, ssize = train_size)
train.keys <- getSample(smplr)
test.keys <- setdiff(data.keys, train.keys)

#Define base modeler
target <- "target"# Regression
regressors <- setdiff(names(data.dt), target)
params <- list(
  "num.trees" = 50,
  "max.depth" = 20
)
base_modeler <- rangerRegressorModeler(params = params,
                                       target = target,
                                       regressors = regressors)

#Instantiate memoryModeler
#memoryModeler parameters
data <- list(data2.dt)
maxData <- 3L
modeler <- memoryRegressorModeler(data = data,
                                  modeler = base_modeler,
                                  maxData = maxData)
#Update stored data
length(modeler$data)
modeler <- updateMemory(object = modeler,
                        data = data3.dt,
                        index = 1)
length(modeler$data)#Length has not changed

}
\seealso{
\code{\link{memoryRegressorModeler}}

\code{\link{addMemory}}

\code{\link{cleanMemory}}
}
